### 2024-07-30 #7057 @editorialbot
[PRE REVIEW]: Simple DFT-D3: Library first implementation of the D3 dispersion correction
**Submitting author:** @awvwgk (<a href="http://orcid.org/0000-0001-7809-771X">Sebastian Ehlert</a>)
**Repository:** https://github.com/dftd3/simple-dftd3
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.0
