### 2024-07-31 #7057 @awvwgk
> **License info**:
>
> 🟡 License found: `GNU Lesser General Public License v3.0` ([Check here](https://opensource.org/licenses) for OSI approval)

The package license currently is [LGPL-3.0-or-later](https://opensource.org/license/lgpl-3-0), it shows yellow rather than a green tick here. Please let me know whether this is an issue.
