### 2024-08-02 #7057 @awvwgk
> On the license. This should not be a problem. However, you repository appears to have two license files there, one for the LGPL license and one for a GPL license. If you use a single file and name it LICENSE (and if the text matches the standard text for the license) then both GitHub and our system should pick this up correctly. So perhaps look into this.

I checked the licensing guidelines from the Free Software Foundation on LGPL, my understanding is that both license (GPL + LGPL addition) files are needed for licensing as LGPL.

> You should also include a copy of the license itself somewhere in the distribution of your program. All programs, whether they are released under the GPL or LGPL, should include [the text version of the GPL](https://www.gnu.org/licenses/gpl.txt). In GNU programs we conventionally put the license in a file called COPYING.

>
> If you are releasing your program under the Lesser GPL, you should also include [the text version of the LGPL](https://www.gnu.org/licenses/lgpl.txt), usually in a file called COPYING.LESSER. Please note that, since the LGPL is a set of additional permissions on top of the GPL, it's crucial to include both licenses so users have all the materials they need to understand their rights.
