### 2024-08-12 08f768b
Workarounds for Cray compilers.
Signed-off-by: Federico Pedron <fedekbe@gmail.com>
