### 2024-08-12 f147b26
Workarounds for Cray compilers (#77)
Signed-off-by: Federico Pedron <fedekbe@gmail.com>
