### 2024-08-15 5d26f40
Fixes invalid escape sequences
Uses raw strings to supress the `SyntaxWarning` thrown by python3.12.
