### 2024-08-20 #7057 @zhubonan
Hello @luthaf and @shahmoradi and @nicorenaud 😃, would you like to review this JOSS submission? You can find more about the JOSS review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Let me know if you are happy to review or not. Thanks in advance.
