+++ .github/workflows/wheel.yml
@@ -70,7 +70,7 @@ jobs:
         - ubuntu-latest
         - windows-latest
         - macos-12
-        python: ['38', '39', '310', '311']
+        python: ['38', '39', '310', '311', '312']
 
     defaults:
       run:
@@ -97,7 +97,7 @@ jobs:
           CIBW_BUILD: "*${{ matrix.python }}-*"
           CIBW_BUILD_VERBOSITY: 1
           CIBW_ARCHS_MACOS: x86_64
-          CIBW_ENVIRONMENT_MACOS: CC=gcc-11 CXX=g++-11 FC=gfortran-11 MACOSX_DEPLOYMENT_TARGET=12.0
+          CIBW_ENVIRONMENT_MACOS: CC=gcc-12 CXX=g++-12 FC=gfortran-12 MACOSX_DEPLOYMENT_TARGET=12.0
           CIBW_BEFORE_BUILD_WINDOWS: choco upgrade mingw && pip install delvewheel
           CIBW_REPAIR_WHEEL_COMMAND_WINDOWS: "delvewheel show {wheel} && delvewheel repair -w {dest_dir} {wheel} --no-mangle-all"
 
