+++ .github/workflows/wheel.yml
@@ -68,7 +68,7 @@ jobs:
       matrix:
         os:
         - ubuntu-latest
-        - windows-latest
+        # windows-latest
         - macos-12
         python: ['38', '39', '310', '311', '312']
 
