### 2024-09-02 5024500
Fix python3.12 build failure (#79)
* See https://cffi.readthedocs.io/en/stable/whatsnew.html#v1-16-0
* Replace the use of deprecated numpy random calls
* Upgrade GCC on MacOS
* Add setuptools in meson as dependency
* Skip Windows wheel build

Co-authored-by: Sebastian Ehlert <28669218+awvwgk@users.noreply.github.com>
