### 2024-09-03 #7169 @editorialbot
[REVIEW]: Simple DFT-D3: Library first implementation of the D3 dispersion correction
**Submitting author:** @awvwgk (<a href="http://orcid.org/0000-0001-7809-771X">Sebastian Ehlert</a>)
**Repository:** https://github.com/dftd3/simple-dftd3
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.2.1
