+++ python/README.rst
@@ -222,7 +222,7 @@ Now you are ready to use ``dftd3``, check if you can import it with
    >>> import dftd3
    >>> from dftd3.libdftd3 import get_api_version
    >>> get_api_version()
-   '1.1.0'
+   '1.1.1'
 
 
 Building the extension module
