+++ .github/workflows/wheel.yml
@@ -68,9 +68,10 @@ jobs:
       matrix:
         os:
         - ubuntu-latest
-        # windows-latest
+        - windows-latest
         - macos-12
-        python: ['38', '39', '310', '311', '312']
+        - macos-latest
+        python: ['38', '39', '310', '311', '312', '313']
 
     defaults:
       run:
@@ -83,7 +84,7 @@ jobs:
           python-version: '3.x'
 
       - name: Install cibuildwheel
-        run: python -m pip install cibuildwheel==2.18.1
+        run: python -m pip install cibuildwheel==2.21.1
 
       - name: Download sdist
         uses: actions/download-artifact@v4
@@ -96,8 +97,11 @@ jobs:
           CIBW_ARCHS: auto64
           CIBW_BUILD: "*${{ matrix.python }}-*"
           CIBW_BUILD_VERBOSITY: 1
-          CIBW_ARCHS_MACOS: x86_64
-          CIBW_ENVIRONMENT_MACOS: CC=gcc-12 CXX=g++-12 FC=gfortran-12 MACOSX_DEPLOYMENT_TARGET=12.0
+          CIBW_ARCHS_MACOS: ${{ matrix.os == 'macos-latest' && 'arm64' || 'x86_64' }}
+          CIBW_ENVIRONMENT_MACOS: >
+             CC=gcc-14 CXX=g++-14 FC=gfortran-14
+             MACOSX_DEPLOYMENT_TARGET=${{ matrix.os == 'macos-latest' && '14.0' || '12.0' }}
+          CIBW_BEFORE_ALL_MACOS: brew install gcc@14
           CIBW_BEFORE_BUILD_WINDOWS: choco upgrade mingw && pip install delvewheel
           CIBW_REPAIR_WHEEL_COMMAND_WINDOWS: "delvewheel show {wheel} && delvewheel repair -w {dest_dir} {wheel} --no-mangle-all"
 
