### 2024-09-19 4a4a5a0
Exclude Python versions not supported
