### 2024-09-19 a6fa0bb
Remove Python 3.13 from CI test
