+++ .github/workflows/build.yml
@@ -242,7 +242,7 @@ jobs:
       matrix:
         os: [ubuntu-latest]
         gcc_v: [10]
-        python_v: ['3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
+        python_v: ['3.8', '3.9', '3.10', '3.11', '3.12']
 
     env:
       FC: gfortran
