### 2024-09-19 c8666f7
Keep testing with oldest meson version
