### 2024-09-19 204bd2f
Fixes invalid escape sequences (#78)
Uses raw strings to supress the `SyntaxWarning` thrown by python3.12

Co-authored-by: Sebastian Ehlert <28669218+awvwgk@users.noreply.github.com>
