+++ assets/ci/build-env.yaml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
 dependencies:
   - meson 0.58.0
-  - fpm 0.3.0
+  - fpm
   - cmake
   - ninja
   - gcovr
