+++ python/mesonpep517.toml
@@ -4,7 +4,7 @@ build-backend = "mesonpy"
 
 [project]
 name = "dftd3"
-version = "1.1.0"
+version = "1.1.1"
 description = "Python API of the DFT-D3 project"
 readme = "README.rst"
 license.text = "LGPL-3.0-or-later"
@@ -14,11 +14,13 @@ classifiers = [
   "Intended Audience :: Science/Research",
   "Programming Language :: Fortran",
   "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.7",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
   "Topic :: Scientific/Engineering :: Chemistry",
   "Topic :: Scientific/Engineering :: Physics",
 ]
