+++ python/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = dftd3
-version = 1.1.0
+version = 1.1.1
 desciption = Python API of the DFT-D3 project
 long_desciption = file: README.rst
 long_description_content_type = text/x-rst
