### 2024-09-22 #7169 @shahmoradi
Thank you, @awvwgk, for this submission. It represents a notable amount of work and merits publication in JOSS. However, I think this software can still improve a few issues.
1. I could not readily find standalone examples to examine the package's functionality. While I see code snippets as examples in the documentation, snippets are often tricky to put together and use. Having fully functional concrete examples with the relevant input data and build scripts would be a great addition to this project, making it more accessible and user-friendly.
2. The tests run fine. Are there any coverage reports that document the code coverage through these tests?
3. Does the package have a CI setup?
4. While the multiple-build approaches are admirable and relatively comprehensive, some non-default build systems (such as Ninja) may be challenging for some users. Could you add a brief set of instructions to the installation documents explaining how to install the Ninja build system? (I am curious to know if there is an advantage to using Ninja in this project rather than the popular make, for example?)
