### 2024-09-22 #7169 @awvwgk
@shahmoradi, thank you for your comments.

> Having fully functional concrete examples with the relevant input data and build scripts would be a great addition to this project, making it more accessible and user-friendly.

That is a good point, for the Fortran and C API one full tutorial in the docs could be quite useful. Could you open this as an issue in the repository?

> Are there any coverage reports that document the code coverage through these tests?

We get coverage reports via gcc with gcovr for the Fortran library and C API and via pytest-cov for the Python API. The reports are uploaded to codecov at https://app.codecov.io/gh/dftd3/simple-dftd3

> Does the package have a CI setup?

The continuous integration setup is done via GitHub actions the main CI workflow test building via CMake, meson and fpm with gcc and Intel compilers as well as the build of the Python API. See https://github.com/dftd3/simple-dftd3/actions/workflows/build.yml
Additionally, there is an extensive workflow for testing Python builds on all available platforms on GHA (Linux, OSX/x86_64, OSX/arm64, Windows).

> Could you add a brief set of instructions to the installation documents explaining how to install the Ninja build system? (I am curious to know if there is an advantage to using Ninja in this project rather than the popular make, for example?)

Make is not always available on all platforms and ninja does provide a decent cross platform support which makes it preferable on platforms like Windows. Let's open an issue regarding the documentation of the build systems in the repo and I will address this soon.
