### 2024-09-23 a8ea30b
Dump values in the format as in tests
