+++ src/dftd3/param.f90
@@ -62,8 +62,9 @@ module dftd3_param
          & p_m08hx_df, p_m11l_df, p_mn15l_df, p_pwp_df, p_r2scanh_df, p_r2scan0_df, &
          & p_r2scan50_df, p_b973c_df, p_dm21_df, p_dm21m_df, p_dm21mc_df, p_dm21mu_df, &
          & p_dsdpbep86_df, p_dsdpbeb95_df, p_dsdpbe_df, p_dodscan66_df, &
-         & p_revdsdblyp_df, p_revdsdpbep86_df, p_revdsdpbeb95_df, p_revdsdpbe_df, & 
-         & p_revdodblyp_df, p_revdodpbep86_df, p_revdodpbeb95_df, p_revdodpbe_df
+         & p_revdsdblyp_df, p_revdsdpbep86_df, p_revdsdpbeb95_df, p_revdsdpbe_df, &
+         & p_revdodblyp_df, p_revdodpbep86_df, p_revdodpbeb95_df, p_revdodpbe_df, &
+         & p_pw91_df
    end enum
 
 contains
@@ -178,6 +179,7 @@ function get_method_id(method) result(id)
    case("pwp", "pw91p86"); id = p_pwp_df
    case("pw1pw"); id = p_pw1pw_df
    case("pw6b95"); id = p_pw6b95_df
+   case("pw91"); id = p_pw91_df
    case("pwb6k"); id = p_pwb6k_df
    case("pwgga"); id = p_pwgga_df
    case("pwpb95"); id = p_pwpb95_df
@@ -322,6 +324,8 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(a1=0.0000_wp, s8=0.2804_wp, a2=6.5745_wp, s6=0.750_wp)
    case(p_pwpb95_df)
       param = d3_param(a1=0.0000_wp, s8=0.2904_wp, a2=7.3141_wp, s6=0.820_wp)
+   case(p_pw91_df)
+      param = d3_param(a1=0.6319_wp, s8=1.9598_wp, a2=4.5718_wp)
    case(p_hf_mixed_df)
       param = d3_param(a1=0.5607_wp, s8=3.9027_wp, a2=4.5622_wp)
    case(p_hf_sv_df)
