### 2024-09-23 d89e2ed
Add tests for scsdrpa75, optscsdrpa75, dsdpbedrpa75, dsdpbeb86drpa75
