### 2024-09-23 4f9b4fa
Add original refs for new D3(BJ) parametrizations
