### 2024-09-23 dftd3/simple-dftd3#95 @AlexBuccheri
Top-Level Documentation on Building and Linking with CMake
In response to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7169)

**Configuration Step**

It would be useful to inform the user of how to point to preinstalled versions of external dependencies with cmake/meson, particularly as users of the Grimme-lab suite of programs may have some installed already. I.e passing `mctc-lib_DIR`, `mstore_DIR`, `toml-f_DIR`, `test-drive_DIR` to cmake, or specifying search paths via `CMAKE_PREFIX_PATH`.

It's a trivial point, but it's also worth noting that the initial configure step will fail if one is offline if cmake has to fall back to Fetch_content. I _assume_ this is also true with meson.

**Building from Source**

CMAKE tells the user:

```cmake
-- Performing Test WITH_QP
-- Performing Test WITH_QP - Success
-- Performing Test WITH_XDP
-- Performing Test WITH_XDP - Failed
```

Although it's guessable that `QP` is quad precision and `XDP` is extended double precision,
I still have to dig through the source to confirm.

Suggestion: A brief reference to this in the documentation would be informative, and something (either in the configure step or the docs) to indicate that these preprocessor variables belong test drive.

**Running the Tests**

One could change the command reference from

```shell
pushd _build && ctest && popd
```

to the more idiomatic:

```shell
ctest --test-dir ./_build
```
