### 2024-09-23 dftd3/simple-dftd3#96 @AlexBuccheri
Top-level README Instructions for Building Docs
In response to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7169)

The top-level README references that ford can be used to build the docs, but does not provide instructions for building. It would be helpful to include:

```shell
# From the project root:
pip install ford
ford docs.md
```
