### 2024-09-23 dftd3/simple-dftd3#97 @AlexBuccheri
Top-level README: Running simple-dftd3 from the CLI Instructions Fail
In response to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7169).

Running the executable with:

```shell
_build/app/s-dftd3 --bj pbe0 --atm --json --noedisp --grad app/example.xyz
```

didn't work for me. I had to specify a file name for the output of the gradient evaluation. I assume the CLI can't have the input file name follow an an option with an optional file name arg, as this is ambiguous. For example, if I change the order to:

```shell
_build/app/s-dftd3 --bj pbe0 --atm --json --grad --noedisp app/example.xyz
```
then the defaults for the JSON and gradient file outputs are picked up. If however I follow either of them with the input structure, it gives an error.

One could note this in the docs, or modify the CLI to do some extra checking as JSON, gradient and input files can be distinguished from their extensions.
