### 2024-09-23 #7169 @AlexBuccheri
Hi @awvwgk, I've done a first pass on the documentation and build process of the project. In response to this, I've opened issues [95](https://github.com/dftd3/simple-dftd3/issues/95), [96](https://github.com/dftd3/simple-dftd3/issues/96) and [97](https://github.com/dftd3/simple-dftd3/issues/97), as well as commented on the two by @shahmoradi

I second their point w.r.t. the need for some visible examples of linking against s-dftd3 for fortran, C and python.
I am yet to try the conda build, but will do so once a simple python example has been added.

The meson build failed for me on Mac Sonoma:

```
config/meson.build:42:12: ERROR: Dependency "OpenMP" not found, tried system
```

but I haven't opened an issue as I checked my aliases and they look like gfortran 14.1.0, GCC and g++ Apple clang version 15.0.0 (clang-1500.3.9.4).

I'm wondering if it's a clang issue, and intend to revisit this as I see the CI is running a mac build no problem. However, one might consider updating the CI to run with the newest GCC on mac, as I see the rest are covered by Ubuntu builds. One might also consider updating intel to run with ifx.
