### 2024-09-23 #7169 @awvwgk
> g++ Apple clang version 15.0.0 (clang-1500.3.9.4).
>
> I'm wondering if it's a clang issue, and intend to revisit this as I see the CI is running a mac build no problem.

This issue is specific to Apple clang only, since it - does not provide OpenMP you either need to use a C compiler with OpenMP enabled or disable OpenMP. Maybe since this problem on OSX has been around the build system could just disable OpenMP for Apple clang by default.
