### 2024-09-23 #7169 @foxtran
For MacOS GCC can be installed and then there is no issues with OpenMP. I have compiled simple-dftd3 with the following commands:
```
cmake -Bbuild -DCMAKE_C_COMPILER=gcc-13 -DCMAKE_Fortran_COMPILER=gfortran-13 -DCMAKE_BUILD_TYPE=Release
cd build
make -j 8
```

It reports that OpenMP is found:
```
-- Found OpenMP_Fortran: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5")
```
