### 2024-09-23 dftd3/simple-dftd3#97 @AlexBuccheri
```
[Error] Insufficient arguments provided
```

which is why I assume  `app/example.xyz` is getting parsed as the file arg for ` --grad`
