### 2024-09-23 #7169 @AlexBuccheri
> For MacOS GCC can be installed and then there is no issues with OpenMP....

Indeed, I already built and tested with cmake. cmake picked up my brew packages without specifying, meson did not. Building with:

```shell
export FC=/opt/homebrew/bin/gfortran-14
export CC=/opt/homebrew/bin/gcc-14
export CXX=/opt/homebrew/bin/g++-14
meson setup _build -Dc_args="-isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk" -Dcpp_args="-isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk"
```

I get the gfortran error:

```shell
[151/217] Compiling Fortran object libs-dftd3.1.dylib.p/src_dftd3_damping_rational.f90.o
FAILED: libs-dftd3.1.dylib.p/src_dftd3_damping_rational.f90.o libs-dftd3.1.dylib.p/dftd3_damping_rational.mod
/opt/homebrew/bin/gfortran-14 -Ilibs-dftd3.1.dylib.p -I. -I.. -Isubprojects/mctc-lib/libmctc-lib.a.p -Isubprojects/json-fortran-8.2.5/libjsonfortran.a.p -fdiagnostics-color=always -Wall -O2 -g -ffree-line-length-none -fbacktrace -fopenmp -fopenmp -Jlibs-dftd3.1.dylib.p -o libs-dftd3.1.dylib.p/src_dftd3_damping_rational.f90.o -c ../src/dftd3/damping/rational.f90
../src/dftd3/damping/rational.f90:68:11:

   68 |    self%s6 = param%s6
      |           1
Error: 's6' at (1) is not a member of the 'damping_param' structure
../src/dftd3/damping/rational.f90:69:11:

   69 |    self%s8 = param%s8
      |           1
Error: 's8' at (1) is not a member of the 'damping_param' structure
../src/dftd3/damping/rational.f90:70:11:

   70 |    self%s9 = param%s9
      |           1
Error: 's9' at (1) is not a member of the 'damping_param' structure
../src/dftd3/damping/rational.f90:71:11:
```

Looks weird to me that gfortran complains when I build with meson but not cmake (unless I'm really picking up different compilers with cmake - I'll check). However, `new_rational_damping` is not type-bound, and the `damping_param` base class is abstract. You could just change the arg declaration to:

` type(rational_damping_param), intent(out) :: self`
