### 2024-09-23 #7169 @AlexBuccheri
> > g++ Apple clang version 15.0.0 (clang-1500.3.9.4).
> > I'm wondering if it's a clang issue, and intend to revisit this as I see the CI is running a mac build no problem.
>
> This issue is specific to Apple clang only, since it does not provide OpenMP. You either need to use a C compiler with OpenMP enabled or disable OpenMP. Maybe since this problem on OSX has been around the build system could just disable OpenMP for Apple clang by default.

Ah, thanks for confirming. Yes, if you're interested in having meson/cmake disable OMP and warn the user if clang is picked up/selected, this would lead to a smoother experience. However, I wouldn't insist on this, as I doubt anyone is running production calcs locally on mac. Adding a note in the build instructions would be fine with me.
