+++ README.md
@@ -7,49 +7,43 @@
 [![docs](https://github.com/dftd3/simple-dftd3/actions/workflows/docs.yml/badge.svg)](https://dftd3.github.io/simple-dftd3/)
 [![codecov](https://codecov.io/gh/dftd3/simple-dftd3/branch/main/graph/badge.svg)](https://codecov.io/gh/dftd3/simple-dftd3)
 
-A simple drop-in replacement for ``dftd3``.
-
-This program provides a small and easy to use implementation of the DFT-D3
-dispersion correction
+This package provides a library first implementation of the DFT-D3 dispersion correction
 (see [*JCP* **132**, 154104 (2010)](https://dx.doi.org/10.1063/1.3382344)
 and [*JCC* **32**, 1456 (2011)](https://dx.doi.org/10.1002/jcc.21759) for details).
-
-It is mostly based on the [`dftd4`](https://github.com/dftd4/dftd4) program and
-borrows one or two ideas from the implementation in [`ased3`](https://github.com/ehermes/ased3).
+Usable via the command line interface of the [`s-dftd3` executable](man/s-dftd3.1.adoc),
+in Fortran via the [`dftd3` module](https://dftd3.readthedocs.io/en/latest/api/fortran.html),
+with the C interface by including the [`dftd3.h` header](https://dftd3.readthedocs.io/en/latest/api/c.html),
+or in Python by using the [`dftd3` module](https://dftd3.readthedocs.io/en/latest/api/python.html).
 
 
 ## Installation
 
+A full guide for installing this package or building from source can be found in the
+[installation guide](https://dftd3.readthedocs.io/en/latest/installation.html).
 
 ### Conda package
 
 [![Conda Version](https://img.shields.io/conda/vn/conda-forge/simple-dftd3.svg?label=simple-dftd3)](https://anaconda.org/conda-forge/simple-dftd3)
 [![Conda Version](https://img.shields.io/conda/vn/conda-forge/dftd3-python.svg?label=dftd3-python)](https://anaconda.org/conda-forge/dftd3-python)
 
-This project is packaged for the *conda* package manager and available on the *conda-forge* channel.
-To install the *conda* package manager we recommend the [miniforge](https://github.com/conda-forge/miniforge/releases) installer.
-If the *conda-forge* channel is not yet enabled, add it to your channels with
+The preferred way of installing this package is from the *conda-forge* distribution via the *mamba* package manager.
+To install the *mamba* package manager the [miniforge](https://github.com/conda-forge/miniforge/releases) installer can be used.
+If you already have the *mamba* package manager installed add the *conda-forge* channel if it is not yet enabled:
 
 ```
-conda config --add channels conda-forge
+mamba config --add channels conda-forge
 ```
 
 Once the *conda-forge* channel has been enabled, this project can be installed with:
 
 ```
-conda install simple-dftd3
+mamba install simple-dftd3
 ```
 
 If you want to enable the Python API as well install
 
 ```
-conda install dftd3-python
-```
-
-It is possible to list all of the versions available on your platform with:
-
-```
-conda search simple-dftd3 --channel conda-forge
+mamba install dftd3-python
 ```
 
 Now you are ready to use ``s-dftd3``.
@@ -57,33 +51,44 @@ Now you are ready to use ``s-dftd3``.
 
 ### Building from Source
 
-To build this project from the source code in this repository you need to have
-a Fortran compiler supporting Fortran 2008 and one of the supported build systems:
-- [meson](https://mesonbuild.com) version 0.55 or newer, with
+To build this project from the source code in this repository you need a Fortran compiler supporting
[…]
