### 2024-09-23 dftd3/simple-dftd3#97 @awvwgk
I think the better solution is to update the README to use `--` to stop parsing command line arguments:

````
_build/app/s-dftd3 --bj pbe0 --atm --json --noedisp --grad -- app/example.xyz
````

This example should serve as a more complex use case of the command line interface with many options. I would rather not change the behavior of the `--grad` option to make this example work.
