### 2024-09-23 #7169 @awvwgk
> I get the gfortran error:

Could you open an issue at the repository for this?
