### 2024-09-23 #7169 @awvwgk
> CMake is using the same gfortran. Stupid question, but when is gcc called? Building with cmake, using the default flags, looks like it only compiles fortran code.

The C compiler is only used to compile the C API test, the core library is pure Fortran and does not need a C compiler.
