### 2024-09-23 #7169 @AlexBuccheri
> > I get the gfortran error:
>
> Could you open an issue at the repository for this?

Yes, opened as [issue 101](https://github.com/dftd3/simple-dftd3/issues/101)
