### 2024-09-23 5269886
Add instructions for running ford to README
