+++ docs.md
@@ -16,6 +16,7 @@ predocmark: >
 source: true
 graph: false
 sort: alpha
+preprocess: false
 print_creation_date: true
 extra_mods: iso_fortran_env:https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
             mctc_io:https://grimme-lab.github.io/mctc-lib/modules/mctc_io.html
