### 2024-09-23 dftd3/simple-dftd3#101 @awvwgk
I am somewhat confused by this error, checking the subroutine `new_rational_damping` I can't see the source for this error:

https://github.com/dftd3/simple-dftd3/blob/1f4e4591f4d7b2f93ee4fc5e39d7b2617d497ae3/src/dftd3/damping/rational.f90#L59-L75
