### 2024-09-23 #7169 @AlexBuccheri
> > CMake is using the same gfortran. Stupid question, but when is gcc called? Building with cmake, using the default flags, looks like it only compiles fortran code.
>
> The C compiler is only used to compile the C API test, the core library is pure Fortran and does not need a C compiler.

This is what I assumed, but this doesn't happen by default with meson or cmake builds? I don't recall seeing a description of how to build the C API in the top-level README (only the python API via meson).
