### 2024-09-23 dftd3/simple-dftd3#101 @AlexBuccheri
Looking at version control, I also can't... I must have changed this when trying to compile a fortran example from the docs and then modified the example. Then fixed the bug I introduced when returning back to the meson build. Apologies about that, I'll shut the issue.
