+++ src/dftd3/param.f90
@@ -463,15 +463,15 @@ subroutine get_rational_damping(param, method, error, s9)
    case(p_revdodpbe_df)
       param = d3_param(s6=0.6067_wp, a1=0.0_wp, s8=0.0_wp, a2=5.5_wp)
    case(p_drpa75_df)
-      param = d3_param(s6=0.3754_wp, a1=0.0_wp, s8=0.0_wp, a2 = 4.5048_wp)
+      param = d3_param(s6=0.3754_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5048_wp)
    case(p_scs_drpa75_df)
-      param = d3_param(s6=0.2528_wp, a1=0.0_wp, s8=0.0_wp, a2 = 4.5050_wp)
+      param = d3_param(s6=0.2528_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    case(p_optscs_drpa75_df)
-      param = d3_param(s6=0.2546_wp, a1=0.0_wp, s8=0.0_wp, a2 = 4.5050_wp)
+      param = d3_param(s6=0.2546_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    case(p_dsd_pbe_drpa75_df)
-      param = d3_param(s6=0.3223_wp, a1=0.0_wp, s8=0.0_wp, a2 = 4.5050_wp)
+      param = d3_param(s6=0.3223_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    case(p_dsd_pbep86_drpa75_df)
-      param = d3_param(s6=0.3012_wp, a1=0.0_wp, s8=0.0_wp, a2 = 4.5050_wp)
+      param = d3_param(s6=0.3012_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    end select
 
    if (present(s9)) then
