### 2024-09-23 #7169 @awvwgk
> This is what I assumed, but this doesn't happen by default with meson or cmake builds? I don't recall seeing a description of how to build the C API in the top-level README (only the python API via meson).

The C API is completely defined in Fortran, no C compiler is necessary. There is a C header file, however this is only installed. The C compiler is used in one of the integration tests for https://github.com/dftd3/simple-dftd3/blob/main/test/api/api-test.c to test whether the API works when used in C.
