+++ src/dftd3/param.f90
@@ -66,7 +66,7 @@ module dftd3_param
          & p_revdodblyp_df, p_revdodpbep86_df, p_revdodpbeb95_df, p_revdodpbe_df, &
          & p_pw91_df, p_drpa75_df, p_scs_drpa75_df, p_optscs_drpa75_df, &
          & p_dsd_pbe_drpa75_df, p_dsd_pbep86_drpa75_df, &
-         & p_dsdpbep86_km11_df
+         & p_dsdpbep86_km11_df, p_drpa75_mrcc_df, p_scs_drpa75_mrcc_df
    end enum
 
 contains
@@ -120,6 +120,7 @@ function get_method_id(method) result(id)
    case("dm21mc"); id = p_dm21mc_df
    case("dm21mu"); id = p_dm21mu_df
    case("drpa75"); id = p_drpa75_df
+   case("drpa75_mrcc"); id = p_drpa75_mrcc_df
    case("dsdblyp"); id = p_dsdblyp_df
    case("dsdblypfc"); id = p_dsdblypfc_df
    case("dsdpbe"); id = p_dsdpbe_df
@@ -214,6 +215,7 @@ function get_method_id(method) result(id)
    case("rscan"); id = p_rscan_df
    case("scan"); id = p_scan_df
    case("scsdrpa75"); id = p_scs_drpa75_df
+   case("scsdrpa75_mrcc"); id = p_scs_drpa75_mrcc_df
    case("slaterdiracexchange"); id = p_slaterdiracexchange_df
    case("sogga11x"); id = p_sogga11x_df
    case("ssb"); id = p_ssb_df
@@ -476,6 +478,10 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.3012_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    case(p_dsdpbep86_km11_df)
       param = d3_param(s6=0.418_wp, a1=0.0_wp, s8=0.0_wp, a2=5.65_wp)
+   case(p_drpa75_mrcc_df)
+      param = d3_param(s6=0.375_wp, a1=0.0_wp, s8=0.0_wp, a2=4.505_wp)
+   case(p_scs_drpa75_mrcc_df)
+      param = d3_param(s6=0.375_wp, a1=0.0_wp, s8=0.0_wp, a2=4.505_wp)
    end select
 
    if (present(s9)) then
