+++ assets/parameters.toml
@@ -474,3 +474,11 @@ d3.bj = {s6=0.3012, a1=0.0, s8=0.0, a2=4.5050, doi="10.1021/acs.jpca.1c01295"}
 
 [parameter.dsdpbep86_km11]
 d3.bj = {s6=0.418, a1=0.0, s8=0.0, a2=5.65, doi="10.1039/c1cp22592h"}
+
+# dRPA75-D3(BJ) from MRCC source code
+[parameter.drpa75_mrcc]
+d3.bj = {s6=0.375, a1=0.0, s8=0.0, a2=4.505}
+
+# SCS-dRPA75-D3(BJ) from MRCC source code
+[parameter.scsdrpa75_mrcc]
+d3.bj = {s6=0.375, a1=0.0, s8=0.0, a2=4.505}
