### 2024-09-24 a5ac4fd
Add DSD-S**-D3(BJ) from 10.1002/jcc.23391
