### 2024-09-24 4b9ce1b
Add DSD-B**-D3(BJ) from 10.1002/jcc.23391
