+++ assets/parameters.toml
@@ -494,3 +494,21 @@ d3.bj = {s6=0.30, a1=0.0, s8=0.0, a2=5.6, doi="10.1002/jcc.23391"}
 
 [parameter.dsdspbe]
 d3.bj = {s6=0.40, a1=0.0, s8=0.0, a2=6.0, doi="10.1002/jcc.23391"}
+
+[parameter.dsdbvwn5]
+d3.bj = {s6=0.61, a1=0.0, s8=0.0, a2=5.2, doi="10.1002/jcc.23391"}
+
+[parameter.dsdblyp_km13]
+d3.bj = {s6=0.57, a1=0.0, s8=0.0, a2=5.4, doi="10.1002/jcc.23391"}
+
+[parameter.dsdbpbe]
+d3.bj = {s6=1.22, a1=0.0, s8=0.0, a2=6.6, doi="10.1002/jcc.23391"}
+
+[parameter.dsdbp86]
+d3.bj = {s6=0.76, a1=0.0, s8=0.0, a2=6.0, doi="10.1002/jcc.23391"}
+
+[parameter.dsdbpw91]
+d3.bj = {s6=1.14, a1=0.0, s8=0.0, a2=6.5, doi="10.1002/jcc.23391"}
+
+[parameter.dsdbb95]
+d3.bj = {s6=1.02, a1=0.0, s8=0.0, a2=6.8, doi="10.1002/jcc.23391"}
