+++ assets/parameters.toml
@@ -100,15 +100,6 @@ d3.zero = {rs6=1.569, s8=0.705, s6=0.5}
 [parameter.dsdblypfc]
 d3.bj = {a1=0.0009, s8=0.2112, a2=5.9807, s6=0.50}
 
-[parameter.dsdpbep86]
-d3.bj = {s6=0.48, a1=0.0, s8=0.0, a2=5.6, doi="10.1002/jcc.23391"}
-
-[parameter.dsdpbeb95]
-d3.bj = {s6=0.61, a1=0.0, s8=0.0, a2=6.2, doi="10.1002/jcc.23391"}
-
-[parameter.dsdpbe]
-d3.bj = {s6=0.78, a1=0.0, s8=0.0, a2=6.1, doi="10.1002/jcc.23391"}
-
 [parameter.dodscan66]
 d3.bj = {s6=0.3152, a1=0.0, s8=0.0, a2=5.75, doi="10.1021/acs.jpca.9b03157"}
 
@@ -512,3 +503,21 @@ d3.bj = {s6=1.14, a1=0.0, s8=0.0, a2=6.5, doi="10.1002/jcc.23391"}
 
 [parameter.dsdbb95]
 d3.bj = {s6=1.02, a1=0.0, s8=0.0, a2=6.8, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbevwn5]
+d3.bj = {s6=0.54, a1=0.0, s8=0.0, a2=5.1, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbelyp]
+d3.bj = {s6=0.43, a1=0.0, s8=0.0, a2=5.2, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbe]
+d3.bj = {s6=0.78, a1=0.0, s8=0.0, a2=6.1, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbep86]
+d3.bj = {s6=0.48, a1=0.0, s8=0.0, a2=5.6, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbepw91]
+d3.bj = {s6=0.73, a1=0.0, s8=0.0, a2=6.0, doi="10.1002/jcc.23391"}
+
+[parameter.dsdpbeb95]
+d3.bj = {s6=0.61, a1=0.0, s8=0.0, a2=6.2, doi="10.1002/jcc.23391"}
