+++ src/dftd3/param.f90
@@ -70,7 +70,7 @@ module dftd3_param
          & p_dsd_svwn5_df, p_dsd_sp86_df, p_dsd_slyp_df, p_dsd_spbe_df, &
          & p_dsd_bvwn5_df, p_dsd_blyp_km13_df, p_dsd_bpbe_df, p_dsd_bp86_df, &
          & p_dsd_bpw91_df, p_dsd_bb95_df, p_dsd_pbevwn5_df, p_dsd_pbelyp_df, &
-         & p_dsd_pbepw91_df
+         & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df
    end enum
 
 contains
@@ -146,6 +146,8 @@ function get_method_id(method) result(id)
    case("dsdpbep86drpa75"); id = p_dsd_pbep86_drpa75_df
    case("dsdpbepw91"); id = p_dsd_pbepw91_df
    case("dsdpbeb95"); id = p_dsdpbeb95_df
+   case("dsdpbehb95"); id = p_dsd_pbehb95_df
+   case("dsdpbehp86"); id = p_dsd_pbehp86_df
    case("dodscan66"); id = p_dodscan66_df
    case("hcth120"); id = p_hcth120_df
    case("hcth407", "hcth/407"); id = p_hcth407_df
@@ -525,6 +527,10 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.73_wp, a1=0.0_wp, s8=0.0_wp, a2=6.0_wp)
    case(p_dsdpbeb95_df)
       param = d3_param(s6=0.61_wp, a1=0.0_wp, s8=0.0_wp, a2=6.2_wp)
+   case(p_dsd_pbehb95_df)
+      param = d3_param(s6=0.58_wp, a1=0.0_wp, s8=0.0_wp, a2=6.2_wp)
+   case(p_dsd_pbehp86_df)
+      param = d3_param(s6=0.46_wp, a1=0.0_wp, s8=0.0_wp, a2=5.6_wp)
    end select
 
    if (present(s9)) then
