### 2024-09-24 8b0dcb6
Add new DSD-mPW**-D3(BJ) from 10.1002/jcc.23391
