+++ src/dftd3/param.f90
@@ -70,7 +70,8 @@ module dftd3_param
          & p_dsd_svwn5_df, p_dsd_sp86_df, p_dsd_slyp_df, p_dsd_spbe_df, &
          & p_dsd_bvwn5_df, p_dsd_blyp_km13_df, p_dsd_bpbe_df, p_dsd_bp86_df, &
          & p_dsd_bpw91_df, p_dsd_bb95_df, p_dsd_pbevwn5_df, p_dsd_pbelyp_df, &
-         & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df
+         & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df, p_dsd_mpwlyp_df, &
+         & p_dsd_mpwpw91_df, p_dsd_mpwp86_df, p_dsd_mpwpbe_df, p_dsd_mpwb95_df
    end enum
 
 contains
@@ -148,6 +149,11 @@ function get_method_id(method) result(id)
    case("dsdpbeb95"); id = p_dsdpbeb95_df
    case("dsdpbehb95"); id = p_dsd_pbehb95_df
    case("dsdpbehp86"); id = p_dsd_pbehp86_df
+   case("dsdmpwlyp"); id = p_dsd_mpwlyp_df
+   case("dsdmpwpw91"); id = p_dsd_mpwpw91_df
+   case("dsdmpwp86"); id = p_dsd_mpwp86_df
+   case("dsdmpwpbe"); id = p_dsd_mpwpbe_df
+   case("dsdmpwb95"); id = p_dsd_mpwb95_df
    case("dodscan66"); id = p_dodscan66_df
    case("hcth120"); id = p_hcth120_df
    case("hcth407", "hcth/407"); id = p_hcth407_df
@@ -531,6 +537,16 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.58_wp, a1=0.0_wp, s8=0.0_wp, a2=6.2_wp)
    case(p_dsd_pbehp86_df)
       param = d3_param(s6=0.46_wp, a1=0.0_wp, s8=0.0_wp, a2=5.6_wp)
+   case(p_dsd_mpwlyp_df)
+      param = d3_param(s6=0.48_wp, a1=0.0_wp, s8=0.0_wp, a2=5.3_wp)
+   case(p_dsd_mpwpw91_df)
+      param = d3_param(s6=0.90_wp, a1=0.0_wp, s8=0.0_wp, a2=6.2_wp)
+   case(p_dsd_mpwp86_df)
+      param = d3_param(s6=0.59_wp, a1=0.0_wp, s8=0.0_wp, a2=5.8_wp)
+   case(p_dsd_mpwpbe_df)
+      param = d3_param(s6=0.96_wp, a1=0.0_wp, s8=0.0_wp, a2=6.3_wp)
+   case(p_dsd_mpwb95_df)
+      param = d3_param(s6=0.82_wp, a1=0.0_wp, s8=0.0_wp, a2=6.6_wp)
    end select
 
    if (present(s9)) then
