### 2024-09-24 96c968e
Add new DSD-HSE**-D3(BJ) from 10.1002/jcc.23391
