+++ src/dftd3/param.f90
@@ -71,7 +71,8 @@ module dftd3_param
          & p_dsd_bvwn5_df, p_dsd_blyp_km13_df, p_dsd_bpbe_df, p_dsd_bp86_df, &
          & p_dsd_bpw91_df, p_dsd_bb95_df, p_dsd_pbevwn5_df, p_dsd_pbelyp_df, &
          & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df, p_dsd_mpwlyp_df, &
-         & p_dsd_mpwpw91_df, p_dsd_mpwp86_df, p_dsd_mpwpbe_df, p_dsd_mpwb95_df
+         & p_dsd_mpwpw91_df, p_dsd_mpwp86_df, p_dsd_mpwpbe_df, p_dsd_mpwb95_df, &
+         & p_dsd_hsepbe_df, p_dsd_hsepw91_df, p_dsd_hsep86_df, p_dsd_hselyp_df
    end enum
 
 contains
@@ -154,6 +155,10 @@ function get_method_id(method) result(id)
    case("dsdmpwp86"); id = p_dsd_mpwp86_df
    case("dsdmpwpbe"); id = p_dsd_mpwpbe_df
    case("dsdmpwb95"); id = p_dsd_mpwb95_df
+   case("dsdhsepbe"); id = p_dsd_hsepbe_df
+   case("dsdhsepw91"); id = p_dsd_hsepw91_df
+   case("dsdhsep86"); id = p_dsd_hsep86_df
+   case("dsdhselyp"); id = p_dsd_hselyp_df
    case("dodscan66"); id = p_dodscan66_df
    case("hcth120"); id = p_hcth120_df
    case("hcth407", "hcth/407"); id = p_hcth407_df
@@ -547,6 +552,14 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.96_wp, a1=0.0_wp, s8=0.0_wp, a2=6.3_wp)
    case(p_dsd_mpwb95_df)
       param = d3_param(s6=0.82_wp, a1=0.0_wp, s8=0.0_wp, a2=6.6_wp)
+   case(p_dsd_hsepbe_df)
+      param = d3_param(s6=0.79_wp, a1=0.0_wp, s8=0.0_wp, a2=6.1_wp)
+   case(p_dsd_hsepw91_df)
+      param = d3_param(s6=0.74_wp, a1=0.0_wp, s8=0.0_wp, a2=6.0_wp)
+   case(p_dsd_hsep86_df)
+      param = d3_param(s6=0.46_wp, a1=0.0_wp, s8=0.0_wp, a2=5.6_wp)
+   case(p_dsd_hselyp_df)
+      param = d3_param(s6=0.40_wp, a1=0.0_wp, s8=0.0_wp, a2=5.2_wp)
    end select
 
    if (present(s9)) then
