+++ src/dftd3/param.f90
@@ -72,7 +72,8 @@ module dftd3_param
          & p_dsd_bpw91_df, p_dsd_bb95_df, p_dsd_pbevwn5_df, p_dsd_pbelyp_df, &
          & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df, p_dsd_mpwlyp_df, &
          & p_dsd_mpwpw91_df, p_dsd_mpwp86_df, p_dsd_mpwpbe_df, p_dsd_mpwb95_df, &
-         & p_dsd_hsepbe_df, p_dsd_hsepw91_df, p_dsd_hsep86_df, p_dsd_hselyp_df
+         & p_dsd_hsepbe_df, p_dsd_hsepw91_df, p_dsd_hsep86_df, p_dsd_hselyp_df, &
+         & p_dsd_tpss_df, p_dsd_tpssb95_df
    end enum
 
 contains
@@ -159,6 +160,8 @@ function get_method_id(method) result(id)
    case("dsdhsepw91"); id = p_dsd_hsepw91_df
    case("dsdhsep86"); id = p_dsd_hsep86_df
    case("dsdhselyp"); id = p_dsd_hselyp_df
+   case("dsdtpss", "dsdtpsstpss"); id = p_dsd_tpss_df
+   case("dsdtpssb95"); id = p_dsd_tpssb95_df
    case("dodscan66"); id = p_dodscan66_df
    case("hcth120"); id = p_hcth120_df
    case("hcth407", "hcth/407"); id = p_hcth407_df
@@ -560,6 +563,10 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.46_wp, a1=0.0_wp, s8=0.0_wp, a2=5.6_wp)
    case(p_dsd_hselyp_df)
       param = d3_param(s6=0.40_wp, a1=0.0_wp, s8=0.0_wp, a2=5.2_wp)
+   case(p_dsd_tpss_df)
+      param = d3_param(s6=0.72_wp, a1=0.0_wp, s8=0.0_wp, a2=6.5_wp)
+   case(p_dsd_tpssb95_df)
+      param = d3_param(s6=0.91_wp, a1=0.0_wp, s8=0.0_wp, a2=7.9_wp)
    end select
 
    if (present(s9)) then
