### 2024-09-24 b78cd41
Add new DSD-O/X**-D3(BJ) from 10.1002/jcc.23391
