+++ assets/parameters.toml
@@ -560,3 +560,12 @@ d3.bj = {s6=0.72, a1=0.0, s8=0.0, a2=6.5, doi="10.1002/jcc.23391"}
 
 [parameter.dsdtpssb95]
 d3.bj = {s6=0.91, a1=0.0, s8=0.0, a2=7.9, doi="10.1002/jcc.23391"}
+
+[parameter.dsdolyp]
+d3.bj = {s6=0.93, a1=0.0, s8=0.0, a2=5.8, doi="10.1002/jcc.23391"}
+
+[parameter.dsdxlyp]
+d3.bj = {s6=0.51, a1=0.0, s8=0.0, a2=5.3, doi="10.1002/jcc.23391"}
+
+[parameter.dsdxb95]
+d3.bj = {s6=0.92, a1=0.0, s8=0.0, a2=6.7, doi="10.1002/jcc.23391"}
