### 2024-09-24 ee1e560
Add new DSD-[B98/BMK/tHCTH/HCTH407]-D3(BJ) from 10.1002/jcc.23391
