### 2024-09-24 d0b3708
Add all DOD-DFT-D3(BJ) from 10.1002/jcc.23391
