### 2024-09-24 f8b1fd0
Put dsd functionals in order
