+++ src/dftd3/param.f90
@@ -163,15 +163,15 @@ function get_method_id(method) result(id)
    case("dsdhsepw91"); id = p_dsd_hsepw91_df
    case("dsdhsep86"); id = p_dsd_hsep86_df
    case("dsdhselyp"); id = p_dsd_hselyp_df
-   case("dsdb98"); id = p_dsd_b98_df
-   case("dsdbmk"); id = p_dsd_bmk_df
-   case("dsdthcth"); id = p_dsd_thcth_df
-   case("dsdhcth407"); id = p_dsd_hcth407_df
    case("dsdtpss", "dsdtpsstpss"); id = p_dsd_tpss_df
    case("dsdtpssb95"); id = p_dsd_tpssb95_df
    case("dsdolyp"); id = p_dsd_olyp_df
    case("dsdxlyp"); id = p_dsd_xlyp_df
    case("dsdxb95"); id = p_dsd_xb95_df
+   case("dsdb98"); id = p_dsd_b98_df
+   case("dsdbmk"); id = p_dsd_bmk_df
+   case("dsdthcth"); id = p_dsd_thcth_df
+   case("dsdhcth407"); id = p_dsd_hcth407_df
    case("dodsvwn5"); id = p_dod_svwn5_df
    case("dodblyp"); id = p_dod_blyp_df
    case("dodpbe", "dodpbepbe"); id = p_dod_pbe_df
