### 2024-09-28 6a20a76
Add citations to parameter file
