+++ assets/parameters.toml
@@ -9,20 +9,20 @@ d3.zerom = {s6=1.0, s9=1.0, rs8=1.0, alp=14.0, damping="mzero"}
 d3.op = {s9=1.0, alp=14.0, damping="optimizedpower"}
 
 [parameter.bp]
-d3.bj = {a1=0.3946, s8=3.2822, a2=4.8516}
+d3.bj = {a1=0.3946, s8=3.2822, a2=4.8516, doi="10.1002/jcc.21759"}
 d3.zero = {rs6=1.139, s8=1.683, doi="10.1063/1.3382344"}
 d3.bjm = {a1=0.821850, s8=3.140281, a2=2.728151, doi="10.1021/acs.jpclett.6b00780"}
 d3.zerom = {rs6=1.233460, s8=1.945174, bet=0.000000, doi="10.1021/acs.jpclett.6b00780"}
 
 [parameter.blyp]
-d3.bj = {a1=0.4298, s8=2.6996, a2=4.2359}
+d3.bj = {a1=0.4298, s8=2.6996, a2=4.2359, doi="10.1002/jcc.21759"}
 d3.zero = {rs6=1.094, s8=1.682, doi="10.1063/1.3382344"}
 d3.bjm = {a1=0.448486, s8=1.875007, a2=3.610679, doi="10.1021/acs.jpclett.6b00780"}
 d3.zerom = {rs6=1.279637, s8=1.841686, bet=0.014370, doi="10.1021/acs.jpclett.6b00780"}
 d3.op = {s6=1.0, s8=1.31867, a1=0.425, a2=3.50, bet=2.0, doi="10.1021/acs.jctc.7b00176"}
 
 [parameter.revpbe]
-d3.bj = {a1=0.5238, s8=2.3550, a2=3.5016}
+d3.bj = {a1=0.5238, s8=2.3550, a2=3.5016, doi="10.1002/jcc.21759"}
 d3.zero = {rs6=0.923, s8=1.010, doi="10.1063/1.3382344"}
 d3.op = {s6=1.0, s8=1.44765, a1=0.600, a2=2.50, bet=0.0, doi="10.1021/acs.jctc.7b00176"}
 
@@ -31,7 +31,7 @@ d3.bj = {a1=0.1820, s8=0.8318, a2=4.0094}
 d3.zero = {rs6=0.872, s8=0.514}
 
 [parameter.b97d]
-d3.bj = {a1=0.5545, s8=2.2609, a2=3.2297}
+d3.bj = {a1=0.5545, s8=2.2609, a2=3.2297, doi="10.1002/jcc.21759"}
 d3.zero = {rs6=0.892, s8=0.909, doi="10.1063/1.3382344"}
 d3.bjm = {a1=0.240184, s8=1.206988, a2=3.864426, doi="10.1021/acs.jpclett.6b00780"}
 d3.zerom = {rs6=1.151808, s8=1.020078, bet=0.035964, doi="10.1021/acs.jpclett.6b00780"}
@@ -42,36 +42,36 @@ d3.bj = {a1=0.37, s8=1.50, a2=4.10, doi="10.1063/1.5012601"}
 d3.zero = {rs6=1.06, s8=1.50, doi="10.1063/1.5012601"}
 
 [parameter.pbe]
-d3.bj = {a1=0.4289, s8=0.7875, a2=4.4407}
+d3.bj = {a1=0.4289, s8=0.7875, a2=4.4407, doi="10.1002/jcc.21759"}
 d3.zero = {rs6=1.217, s8=0.722, doi="10.1063/1.3382344"}
[…]
