### 2024-09-28 950ace5
Add further references
