+++ assets/parameters.toml
@@ -78,8 +78,8 @@ d3.zerom = {rs6=2.077949, s8=0.000081, bet=0.116755, doi="10.1021/acs.jpclett.6b
 d3.op = {s6=0.88290, s8=0.0, a1=0.150, a2=4.750, bet=6.0, doi="10.1021/acs.jctc.7b00176"}
 
 [parameter.hse06]
-d3.bj = {a1=0.383, s8=2.310, a2=5.685}
-d3.zero = {rs6=1.129, s8=0.109}
+d3.bj = {a1=0.383, s8=2.310, a2=5.685, doi="10.1021/jp501237c"}
+d3.zero = {rs6=1.129, s8=0.109, doi="10.1021/jp501237c"}
 
 [parameter.revpbe38]
 d3.bj = {a1=0.4309, s8=1.4760, a2=3.9446, doi="10.1039/c0cp02984j"}
@@ -89,13 +89,13 @@ d3.bj = {a1=0.2076, s8=0.7257, a2=6.3750, doi="10.1039/c0cp02984j"}
 
 [parameter.b2plyp]
 d3.bj = {a1=0.3065, s8=0.9147, a2=5.0570, s6=0.64, doi="10.1039/c0cp02984j"}
-d3.zero = {rs6=1.427, s8=1.022, s6=0.64}
+d3.zero = {rs6=1.427, s8=1.022, s6=0.64, doi="10.1021/ct100466k"}
 d3.bjm = {a1=0.486434, s8=0.672820, a2=3.656466, s6=0.64, doi="10.1021/acs.jpclett.6b00780"}
 d3.zerom = {rs6=1.313134, s8=0.717543, bet=0.016035, s6=0.64, doi="10.1021/acs.jpclett.6b00780"}
 
 [parameter.dsdblyp]
 d3.bj = {a1=0.0000, s8=0.2130, a2=6.0519, s6=0.50, doi="10.1039/c0cp02984j"}
-d3.zero = {rs6=1.569, s8=0.705, s6=0.5, doi="10.1039/c0cp02984j"}
+d3.zero = {rs6=1.569, s8=0.705, s6=0.5, doi="10.1021/ct100466k"}
 
 [parameter.dsdblypfc]
 d3.bj = {a1=0.0009, s8=0.2112, a2=5.9807, s6=0.50, doi="10.1039/c0cp02984j"}
@@ -132,22 +132,22 @@ d3.bj = {a1=0.4870, s8=3.2950, a2=3.5043, doi="10.1039/c0cp02984j"}
 
 [parameter.mpwlyp]
 d3.bj = {a1=0.4831, s8=2.0077, a2=4.5323, doi="10.1039/c0cp02984j"}
-d3.zero = {rs6=1.239, s8=1.098}
+d3.zero = {rs6=1.239, s8=1.098, doi="10.1039/c0cp02984j"}
 
 [parameter.olyp]
 d3.bj = {a1=0.5299, s8=2.6205, a2=2.8065, doi="10.1039/c0cp02984j"}
-d3.zero = {rs6=0.806, s8=1.764}
+d3.zero = {rs6=0.806, s8=1.764, doi="10.1039/c0cp02984j"}
 
 [parameter.pbesol]
 d3.bj = {a1=0.4466, s8=2.9491, a2=6.1742, doi="10.1039/c0cp02984j"}
 
 [parameter.bpbe]
 d3.bj = {a1=0.4567, s8=4.0728, a2=4.3908, doi="10.1039/c0cp02984j"}
-d3.zero = {rs6=1.087, s8=2.033}
+d3.zero = {rs6=1.087, s8=2.033, doi="10.1039/c0cp02984j"}
 
 [parameter.opbe]
[…]
