+++ assets/parameters.toml
@@ -717,7 +717,7 @@ published = 2019-05-28
 
 [citation."10.1073/pnas.1516984112"]  # PW91-D3
 title = "A priori calculations of the free energy of formation from solution of polymorphic self-assembled monolayers"
-author = ["Jeffrey R. Reimers", "Dwi Panduwinata", "Johan Visser" "Maxwell J. Crossley"]
+author = ["Jeffrey R. Reimers", "Dwi Panduwinata", "Johan Visser", "Maxwell J. Crossley"]
 journal = "Proc. Natl. Acad. Sci."
 issue = "45"
 volume = "112"
