+++ assets/parameters.toml
@@ -492,7 +492,7 @@ d3.bj = {s6=0.40, a1=0.0, s8=0.0, a2=6.0, doi="10.1002/jcc.23391"}
 [parameter.dsdbvwn5]
 d3.bj = {s6=0.61, a1=0.0, s8=0.0, a2=5.2, doi="10.1002/jcc.23391"}
 
-[parameter.dsdblyp_km13]
+[parameter.dsdblyp_2013]
 d3.bj = {s6=0.57, a1=0.0, s8=0.0, a2=5.4, doi="10.1002/jcc.23391"}
 
 [parameter.dsdbpbe]
