+++ src/dftd3/param.f90
@@ -520,7 +520,7 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.3223_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
    case(p_dsd_pbep86_drpa75_df)
       param = d3_param(s6=0.3012_wp, a1=0.0_wp, s8=0.0_wp, a2=4.5050_wp)
-   case(p_dsdpbep86_km11_df)
+   case(p_dsdpbep86_2011_df)
       param = d3_param(s6=0.418_wp, a1=0.0_wp, s8=0.0_wp, a2=5.65_wp)
    case(p_drpa75_mrcc_df)
       param = d3_param(s6=0.375_wp, a1=0.0_wp, s8=0.0_wp, a2=4.505_wp)
@@ -536,7 +536,7 @@ subroutine get_rational_damping(param, method, error, s9)
       param = d3_param(s6=0.40_wp, a1=0.0_wp, s8=0.0_wp, a2=6.0_wp)
    case(p_dsd_bvwn5_df)
       param = d3_param(s6=0.61_wp, a1=0.0_wp, s8=0.0_wp, a2=5.2_wp)
-   case(p_dsd_blyp_km13_df)
+   case(p_dsd_blyp_2013_df)
       param = d3_param(s6=0.57_wp, a1=0.0_wp, s8=0.0_wp, a2=5.4_wp)
    case(p_dsd_bpbe_df)
       param = d3_param(s6=1.22_wp, a1=0.0_wp, s8=0.0_wp, a2=6.6_wp)
