### 2024-09-28 29128f2
Remove MRCC specific parameters without reference
