+++ test/unit/test_param.f90
@@ -114,14 +114,14 @@ subroutine test_d3bj_mb01(error)
       & "dm21", "dm21m", "dm21mc", "dm21mu", "dsdpbep86", "dsdpbeb95", "dsdpbe", &
       & "dodscan66", "revdsdblyp", "revdsdpbep86", "revdsdpbeb95", "revdsdpbe", &
       & "revdodblyp", "revdodpbep86", "revdodpbeb95", "revdodpbe", "pw91", "drpa75", &
-      & "scsdrpa75", "optscsdrpa75", "dsdpbedrpa75", "dsdpbep86drpa75", "dsdpbep86_2011", &
-      & "drpa75_mrcc", "scsdrpa75_mrcc", "dsdsvwn5", "dsdsp86", "dsdslyp", "dsdspbe", &
-      & "dsdbvwn5", "dsdblyp_2013", "dsdbpbe", "dsdbp86", "dsdbpw91", "dsdbb95", &
-      & "dsdpbevwn5", "dsdpbelyp", "dsdpbepw91", "dsdpbehb95", "dsdpbehp86", &
-      & "dsdmpwlyp", "dsdmpwpw91", "dsdmpwp86", "dsdmpwpbe", "dsdmpwb95", "dsdhsepbe", &
-      & "dsdhsepw91", "dsdhsep86", "dsdhselyp", "dsdtpss", "dsdtpsstpss", "dsdtpssb95", &
-      & "dsdolyp", "dsdxlyp", "dsdxb95", "dsdb98", "dsdbmk", "dsdthcth", "dsdhcth407", &
-      & "dodsvwn5", "dodblyp", "dodpbe", "dodpbep86", "dodpbeb95", "dodhsep86", "dodpbehb95"]
+      & "scsdrpa75", "optscsdrpa75", "dsdpbedrpa75", "dsdpbep86drpa75", &
+      & "dsdsvwn5", "dsdsp86", "dsdslyp", "dsdspbe", "dsdbvwn5", "dsdblyp_2013", "dsdbpbe", &
+      & "dsdbp86", "dsdbpw91", "dsdbb95", "dsdpbevwn5", "dsdpbelyp", "dsdpbepw91", &
+      & "dsdpbehb95", "dsdpbehp86", "dsdmpwlyp", "dsdmpwpw91", "dsdmpwp86", "dsdmpwpbe", &
+      & "dsdmpwb95", "dsdhsepbe", "dsdhsepw91", "dsdhsep86", "dsdhselyp", "dsdtpss", &
+      & "dsdtpsstpss", "dsdtpssb95", "dsdolyp", "dsdxlyp", "dsdxb95", "dsdb98", "dsdbmk", &
+      & "dsdthcth", "dsdhcth407", "dodsvwn5", "dodblyp", "dodpbe", "dodpbep86", &
+      & "dodpbeb95", "dodhsep86", "dodpbehb95", "dsdpbep86_2011"]
    real(wp), parameter :: ref(*) = [&
       &-2.9551694676908012E-2_wp,-1.6638703086788331E-2_wp,-1.6725877716130381E-2_wp, &
       &-3.3014429592265318E-2_wp,-2.2051435219996540E-2_wp,-3.3481565825316001E-2_wp, &
@@ -162,8 +162,7 @@ subroutine test_d3bj_mb01(error)
       &-2.7103838130103503E-2_wp,-1.6794975103307797E-2_wp,-1.4460579192722246E-2_wp, &
       &-2.1361659109385409E-2_wp,-9.7460845051874183E-3_wp,-2.8391262582808205E-2_wp, &
       &-1.9116077511614916E-2_wp,-1.9252188823010906E-2_wp,-2.4371486479404614E-2_wp, &
[…]
