+++ assets/parameters.toml
@@ -846,7 +846,7 @@ issue = "1"
 volume = "9"
 pages = "263-272"
 year = "2013"
-published = 2012-11-0509
+published = 2012-11-05
 
 [citation."10.1021/jp501237c"]  # HSE06-D3
 title = "DFT-D3 Study of Some Molecular Crystals"
