### 2024-09-28 b60a7cb
Cleanup journal name
