### 2024-09-28 dd8e4a5
Add more D3(BJ) parametrizations (#103)
* Add DSD-PBEP86-D3BJ from 10.1039/c1cp22592h
* Add DSD-S**-D3(BJ) from 10.1002/jcc.23391
* Add DSD-B**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-PBE**-D3(BJ) from 10.1002/jcc.23391; resort previously added
* Add new DSD-PBEh**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-mPW**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-HSE**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-TPSS**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-O/X**-D3(BJ) from 10.1002/jcc.23391
* Add new DSD-[B98/BMK/tHCTH/HCTH407]-D3(BJ) from 10.1002/jcc.23391
* Add all DOD-DFT-D3(BJ) from 1
[…]
