+++ src/dftd3/param.f90
@@ -65,7 +65,17 @@ module dftd3_param
          & p_revdsdblyp_df, p_revdsdpbep86_df, p_revdsdpbeb95_df, p_revdsdpbe_df, &
          & p_revdodblyp_df, p_revdodpbep86_df, p_revdodpbeb95_df, p_revdodpbe_df, &
          & p_pw91_df, p_drpa75_df, p_scs_drpa75_df, p_optscs_drpa75_df, &
-         & p_dsd_pbe_drpa75_df, p_dsd_pbep86_drpa75_df
+         & p_dsd_pbe_drpa75_df, p_dsd_pbep86_drpa75_df, p_dsdpbep86_2011_df, &
+         & p_dsd_svwn5_df, p_dsd_sp86_df, p_dsd_slyp_df, p_dsd_spbe_df, &
+         & p_dsd_bvwn5_df, p_dsd_blyp_2013_df, p_dsd_bpbe_df, p_dsd_bp86_df, &
+         & p_dsd_bpw91_df, p_dsd_bb95_df, p_dsd_pbevwn5_df, p_dsd_pbelyp_df, &
+         & p_dsd_pbepw91_df, p_dsd_pbehb95_df, p_dsd_pbehp86_df, p_dsd_mpwlyp_df, &
+         & p_dsd_mpwpw91_df, p_dsd_mpwp86_df, p_dsd_mpwpbe_df, p_dsd_mpwb95_df, &
+         & p_dsd_hsepbe_df, p_dsd_hsepw91_df, p_dsd_hsep86_df, p_dsd_hselyp_df, &
+         & p_dsd_tpss_df, p_dsd_tpssb95_df, p_dsd_olyp_df, p_dsd_xlyp_df, &
+         & p_dsd_xb95_df, p_dsd_b98_df, p_dsd_bmk_df, p_dsd_thcth_df, &
+         & p_dsd_hcth407_df, p_dod_svwn5_df, p_dod_blyp_df, p_dod_pbe_df, &
+         & p_dod_pbep86_df, p_dod_pbeb95_df, p_dod_hsep86_df, p_dod_pbehb95_df
    end enum
 
 contains
@@ -119,13 +129,54 @@ function get_method_id(method) result(id)
    case("dm21mc"); id = p_dm21mc_df
    case("dm21mu"); id = p_dm21mu_df
    case("drpa75"); id = p_drpa75_df
+   case("dsdsvwn5"); id = p_dsd_svwn5_df
+   case("dsdsp86"); id = p_dsd_sp86_df
+   case("dsdslyp"); id = p_dsd_slyp_df
+   case("dsdspbe"); id = p_dsd_spbe_df
+   case("dsdbvwn5"); id = p_dsd_bvwn5_df
    case("dsdblyp"); id = p_dsdblyp_df
+   case("dsdblyp_2013"); id = p_dsd_blyp_2013_df
    case("dsdblypfc"); id = p_dsdblypfc_df
-   case("dsdpbe"); id = p_dsdpbe_df
+   case("dsdbpbe"); id = p_dsd_bpbe_df
+   case("dsdbp86"); id = p_dsd_bp86_df
+   case("dsdbpw91"); id = p_dsd_bpw91_df
+   case("dsdbb95"); id = p_dsd_bb95_df
+   case("dsdpbevwn5"); id = p_dsd_pbevwn5_df
+   case("dsdpbelyp"); id = p_dsd_pbelyp_df
+   case("dsdpbe", "dsdpbepbe"); id = p_dsdpbe_df
[…]

[…]
