### 2024-09-28 256c6ad
Add citations for damping parameters
