+++ src/dftd3/param.f90
@@ -16,6 +16,7 @@
 
 module dftd3_param
    use mctc_env, only : wp, error_type, fatal_error
+   use dftd3_citation, only : citation_type, author_name, new_citation
    implicit none
 
    public :: d3_param
@@ -34,6 +35,7 @@ module dftd3_param
       real(wp) :: a2 = 5.0_wp
       real(wp) :: alp = 14.0_wp
       real(wp) :: bet = 0.0_wp
+      type(citation_type), allocatable :: citation
    end type d3_param
 
 
@@ -57,7 +59,7 @@ module dftd3_param
          & p_mpw1kcis_df, p_pbeh1pbe_df, p_pbe1kcis_df, p_x3lyp_df, p_o3lyp_df, &
          & p_b97_1_df, p_b97_2_df, p_b98_df, p_hiss_df, p_hse03_df, p_revtpssh_df, &
          & p_revtpss0_df, p_tpss1kcis_df, p_tauhcthhyb_df, p_mn15_df, p_lc_whpbe_df, &
-         & p_mpw2plyp_df, p_m11_df, p_sogga11x_df, p_n12sx_df, p_mn12sx_df, &
+         & p_mpw2plyp_df, p_m11_df, p_sogga11x_df, p_n12sx_df, p_mn12sx_df, p_mn12l_df, &
          & p_ms2_df, p_ms2h_df, p_mpw1lyp_df, p_mpwkcis1k_df, p_pkzb_df, p_n12_df, &
          & p_m08hx_df, p_m11l_df, p_mn15l_df, p_pwp_df, p_r2scanh_df, p_r2scan0_df, &
          & p_r2scan50_df, p_b973c_df, p_dm21_df, p_dm21m_df, p_dm21mc_df, p_dm21mu_df, &
@@ -77,6 +79,34 @@ module dftd3_param
          & p_dsd_hcth407_df, p_dod_svwn5_df, p_dod_blyp_df, p_dod_pbe_df, &
          & p_dod_pbep86_df, p_dod_pbeb95_df, p_dod_hsep86_df, p_dod_pbehb95_df
    end enum
+
+   character(len=*), parameter :: &
+      & doi_dftd3_0 = "10.1063/1.3382344", &
+      & doi_dftd3_bj = "10.1002/jcc.21759", &
+      & doi_dftd3_m = "10.1021/acs.jpclett.6b00780", &
+      & doi_dftd3_op = "10.1021/acs.jctc.7b00176", &
+      & doi_gmtkn30_0 = "10.1021/ct100466k", &
+      & doi_gmtkn30_bj = "10.1039/c0cp02984j", &
+      & doi_gmtkn55 = "10.1039/c7cp04913g", &
+      & doi_dsd = "10.1002/jcc.23391", &
+      & doi_dsdpbep86 = "10.1039/c1cp22592h", &
+      & doi_drpa = "10.1021/acs.jpca.1c01295", &
+      & doi_revdsd = "10.1021/acs.jpca.9b03157", &
+      & doi_pw91_d3 = "10.1073/pnas.1516984112", &
+      & doi_r2scan_d4 = "10.1063/5.0041008", &
+      & doi_scan_d3 = "10.1103/physrevb.94.115144", &
+      & doi_pbeh3c = "10.1063/1.4927476", &
+      & doi_hse3c = "10.1039/c6cp01697a", &
+      & doi_b973c = "10.1063/1.5012601", &
+      & doi_hf3c = "10.1002/jcc.23317", &
+      & doi_gcp = "10.1063/1.3700154", &
[…]
