+++ app/help.f90
@@ -54,6 +54,7 @@ module dftd3_app_help
       "                         attempts to add to Turbomole gradient and gradlatt files"//nl//&
       "   --property            Evaluate dispersion related properties"//nl//&
       "   --pair-resolved       Calculate pairwise representation of dispersion energy"//nl//&
+      "   --citation [file]     Print citation information to file (default: dftd3.bib)"//nl//&
       "-v,--verbose             Show more, can be used multiple times"//nl//&
       "-s,--silent              Show less, use twice to supress all output"//nl//&
       "   --version             Print program version and exit"//nl//&
