+++ src/dftd3/citation.f90
@@ -14,45 +14,97 @@
 ! You should have received a copy of the GNU Lesser General Public License
 ! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
 
+!> Module for handling citation data
 module dftd3_citation
    implicit none
    private
 
-   public :: citation_type, author_name, new_citation
-   public :: format_bibtex
+   public :: citation_type, author_name, new_citation, is_citation_present
+   public :: format_bibtex, get_citation, same_citation
+   public :: doi_dftd3_0, doi_dftd3_bj, doi_dftd3_m, doi_dftd3_op, &
+      & doi_gmtkn30_0, doi_gmtkn30_bj, doi_gmtkn55, doi_dsd, doi_dsdpbep86, &
+      & doi_drpa, doi_revdsd, doi_pw91_d3, doi_r2scan_d4, doi_scan_d3, &
+      & doi_pbeh3c, doi_hse3c, doi_b973c, doi_hf3c, doi_gcp, doi_d3pbc, &
+      & doi_r2scan_hyb, doi_r2scan_dhdf, doi_minnesota_d3, doi_b97m_d3, &
+      & doi_wb97x_d3, doi_hse06_d3
 
-   type :: author_name
+   !> Represents an author to allow creating author lists
+   type :: author_type
+      !> Name of the author
       character(len=:), allocatable :: name
-   end type author_name
+   end type author_type
 
-   interface author_name
-      module procedure :: new_author
-   end interface author_name
-
+   !> Represents a citation
    type :: citation_type
+      !> Title of the publication
       character(len=:), allocatable :: title
-      type(author_name), allocatable :: author(:)
+      !> Authors of the publication
+      type(author_type), allocatable :: author(:)
+      !> Issue of the publication
       character(len=:), allocatable :: issue
+      !> Journal of the publication
       character(len=:), allocatable :: journal
+      !> Volume of the publication
       character(len=:), allocatable :: volume
+      !> Year of the publication
       character(len=:), allocatable :: year
+      !> Page numbers of the publication
       character(len=:), allocatable :: pages
+      !> Digital Object Identifier (DOI) of the publication
       character(len=:), allocatable :: doi
    end type citation_type
 
    character(len=*), parameter :: nl = new_line('a')
 
+   character(len=*), parameter :: &
+      & doi_dftd3_0 = "10.1063/1.3382344", &
+      & doi_dftd3_bj = "10.1002/jcc.21759", &
+      & doi_dftd3_m = "10.1021/acs.jpclett.6b00780", &
+      & doi_dftd3_op = "10.1021/acs.jctc.7b00176", &
+      & doi_gmtkn30_0 = "10.1021/ct100466k", &
+      & doi_gmtkn30_bj = "10.1039/c0cp02984j", &
+      & doi_gmtkn55 = "10.1039/c7cp04913g", &
+      & doi_dsd = "10.1002/jcc.23391", &
+      & doi_dsdpbep86 = "10.1039/c1cp22592h", &
+      & doi_drpa = "10.1021/acs.jpca.1c01295", &
+      & doi_revdsd = "10.1021/acs.jpca.9b03157", &
+      & doi_pw91_d3 = "10.1073/pnas.1516984112", &
+      & doi_r2scan_d4 = "10.1063/5.0041008", &
+      & doi_scan_d3 = "10.1103/physrevb.94.115144", &
+      & doi_pbeh3c = "10.1063/1.4927476", &
+      & doi_hse3c = "10.1039/c6cp01697a", &
[…]

[…]
