### 2024-09-29 56b37f8
Add citation as optional argument
