+++ src/dftd3/param.f90
@@ -1060,7 +1060,7 @@ subroutine get_zero_damping(param, method, error, s9, citation)
    end select
 
    if (.not.allocated(doi)) doi = doi_dftd3_0
-   citation = get_citation(doi)
+   if (present(citation)) citation = get_citation(doi)
 
    if (present(s9)) then
       param%s9 = s9
