### 2024-09-29 202a6ba
Add citations for damping parameters (#104)
