+++ src/dftd3/citation.f90
@@ -1,0 +1,611 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+!> Module for handling citation data
+module dftd3_citation
+   implicit none
+   private
+
+   public :: citation_type, author_name, new_citation, is_citation_present
+   public :: format_bibtex, get_citation, same_citation
+   public :: doi_dftd3_0, doi_dftd3_bj, doi_dftd3_m, doi_dftd3_op, &
+      & doi_gmtkn30_0, doi_gmtkn30_bj, doi_gmtkn55, doi_dsd, doi_dsdpbep86, &
+      & doi_drpa, doi_revdsd, doi_pw91_d3, doi_r2scan_d4, doi_scan_d3, &
+      & doi_pbeh3c, doi_hse3c, doi_b973c, doi_hf3c, doi_gcp, doi_d3pbc, &
+      & doi_r2scan_hyb, doi_r2scan_dhdf, doi_minnesota_d3, doi_b97m_d3, &
+      & doi_wb97x_d3, doi_hse06_d3
+
+   !> Represents an author to allow creating author lists
+   type :: author_type
+      !> Name of the author
+      character(len=:), allocatable :: name
+   end type author_type
+
+   !> Represents a citation
+   type :: citation_type
+      !> Title of the publication
+      character(len=:), allocatable :: title
+      !> Authors of the publication
+      type(author_type), allocatable :: author(:)
+      !> Issue of the publication
+      character(len=:), allocatable :: issue
+      !> Journal of the publication
+      character(len=:), allocatable :: journal
+      !> Volume of the publication
+      character(len=:), allocatable :: volume
+      !> Year of the publication
+      character(len=:), allocatable :: year
+      !> Page numbers of the publication
+      character(len=:), allocatable :: pages
+      !> Digital Object Identifier (DOI) of the publication
+      character(len=:), allocatable :: doi
+   end type citation_type
+
+   character(len=*), parameter :: nl = new_line('a')
+
+   character(len=*), parameter :: &
+      & doi_dftd3_0 = "10.1063/1.3382344", &
+      & doi_dftd3_bj = "10.1002/jcc.21759", &
+      & doi_dftd3_m = "10.1021/acs.jpclett.6b00780", &
+      & doi_dftd3_op = "10.1021/acs.jctc.7b00176", &
+      & doi_gmtkn30_0 = "10.1021/ct100466k", &
+      & doi_gmtkn30_bj = "10.1039/c0cp02984j", &
+      & doi_gmtkn55 = "10.1039/c7cp04913g", &
+      & doi_dsd = "10.1002/jcc.23391", &
+      & doi_dsdpbep86 = "10.1039/c1cp22592h", &
+      & doi_drpa = "10.1021/acs.jpca.1c01295", &
+      & doi_revdsd = "10.1021/acs.jpca.9b03157", &
[…]
