+++ src/dftd3/CMakeLists.txt
@@ -21,6 +21,7 @@ set(dir "${CMAKE_CURRENT_SOURCE_DIR}")
 
 list(
   APPEND srcs
+  "${dir}/citation.f90"
   "${dir}/cutoff.f90"
   "${dir}/damping.f90"
   "${dir}/data.f90"

[…]
