+++ src/atm.f90
@@ -1,0 +1,462 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+module dftd3_damping_atm
+   use mctc_env, only : wp
+   use mctc_io, only : structure_type
+   implicit none
+
+   public :: get_atm_dispersion, get_atm_pairwise_dispersion
+
+
+contains
+
+
+!> Evaluation of the dispersion energy expression
+subroutine get_atm_dispersion(mol, trans, cutoff, s9, rs9, alp, rvdw, c6, dc6dcn, &
+      & energy, dEdcn, gradient, sigma)
+
+   !> Molecular structure data
+   class(structure_type), intent(in) :: mol
+
+   !> Lattice points
+   real(wp), intent(in) :: trans(:, :)
+
+   !> Real space cutoff
+   real(wp), intent(in) :: cutoff
+
+   !> Scaling for dispersion coefficients
+   real(wp), intent(in) :: s9
+
+   !> Scaling for van-der-Waals radii in damping function
+   real(wp), intent(in) :: rs9
+
+   !> Exponent of zero damping function
+   real(wp), intent(in) :: alp
+
+   !> Van-der-Waals radii for all element pairs
+   real(wp), intent(in) :: rvdw(:, :)
+
+   !> C6 coefficients for all atom pairs.
+   real(wp), intent(in) :: c6(:, :)
+
+   !> Derivative of the C6 w.r.t. the coordination number
+   real(wp), intent(in), optional :: dc6dcn(:, :)
+
+   !> Dispersion energy
+   real(wp), intent(inout) :: energy(:)
+
+   !> Derivative of the energy w.r.t. the coordination number
+   real(wp), intent(inout), optional :: dEdcn(:)
+
+   !> Dispersion gradient
+   real(wp), intent(inout), optional :: gradient(:, :)
+
+   !> Dispersion virial
+   real(wp), intent(inout), optional :: sigma(:, :)
+
+   logical :: grad
+
+   if (abs(s9) < epsilon(1.0_wp)) return
+   grad = present(dc6dcn) .and. present(dEdcn) .and. present(gradient) &
+      & .and. present(sigma)
+
+   if (grad) then
+      call get_atm_dispersion_derivs(mol, trans, cutoff, s9, rs9, alp, rvdw, c6, dc6dcn, &
+      & energy, dEdcn, gradient, sigma)
+   else
+      call get_atm_dispersion_energy(mol, trans, cutoff, s9, rs9, alp, rvdw, c6, energy)
+   end if
+
+end subroutine get_atm_dispersion
+
+
+!> Evaluation of the dispersion energy expression
[…]
