+++ src/data.f90
@@ -1,0 +1,27 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+module dftd3_data
+   use dftd3_data_covrad, only : get_covalent_rad
+   use dftd3_data_r4r2, only : get_r4r2_val
+   use dftd3_data_vdwrad, only : get_vdw_rad
+   implicit none
+   private
+
+   public :: get_covalent_rad, get_r4r2_val, get_vdw_rad
+
+
+end module dftd3_data
