+++ src/param.f90
@@ -1,0 +1,1259 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+module dftd3_param
+   use mctc_env, only : wp, error_type, fatal_error
+   use dftd3_citation, only : citation_type, author_name, new_citation, &
+      & get_citation, doi_dftd3_0, doi_dftd3_bj, doi_dftd3_m, doi_dftd3_op, &
+      & doi_gmtkn30_0, doi_gmtkn30_bj, doi_gmtkn55, doi_dsd, doi_dsdpbep86, &
+      & doi_drpa, doi_revdsd, doi_pw91_d3, doi_r2scan_d4, doi_scan_d3, &
+      & doi_pbeh3c, doi_hse3c, doi_b973c, doi_hf3c, doi_gcp, doi_d3pbc, &
+      & doi_r2scan_hyb, doi_r2scan_dhdf, doi_minnesota_d3, doi_b97m_d3, &
+      & doi_wb97x_d3, doi_hse06_d3
+   implicit none
+
+   public :: d3_param
+   public :: get_rational_damping, get_zero_damping
+   public :: get_mrational_damping, get_mzero_damping
+   public :: get_optimizedpower_damping
+
+
+   type :: d3_param
+      real(wp) :: s6 = 1.0_wp
+      real(wp) :: s8 = 1.0_wp
+      real(wp) :: s9 = 0.0_wp
+      real(wp) :: rs6 = 1.0_wp
+      real(wp) :: rs8 = 1.0_wp
+      real(wp) :: a1 = 0.4_wp
+      real(wp) :: a2 = 5.0_wp
+      real(wp) :: alp = 14.0_wp
+      real(wp) :: bet = 0.0_wp
+   end type d3_param
+
+
+   enum, bind(C)
+      enumerator :: p_invalid, &
+         & p_bp_df, p_blyp_df, p_revpbe_df, p_rpbe_df, p_b97d_df, p_pbe_df, &
+         & p_rpw86pbe_df, p_b3lyp_df, p_b3lyp_g_df, p_tpss_df, p_hf_df, p_tpss0_df, &
+         & p_pbe0_df, p_hse06_df, p_revpbe38_df, p_pw6b95_df, p_b2plyp_df, &
+         & p_dsdblyp_df, p_dsdblypfc_df, p_bop_df, p_mpwlyp_df, p_olyp_df, &
+         & p_pbesol_df, p_bpbe_df, p_opbe_df, p_ssb_df, p_revssb_df, p_otpss_df, &
[…]
