### 2024-10-14 dftd3/simple-dftd3#95 @AlexBuccheri
W.r.t. informing the user of which package a specific cmake message comes from, I just found one can do this with `CMAKE_MESSAGE_CONTEXT`. For example in test_drive `CMakeLists.txt`, one could bookend with

```cmake
cmake_minimum_required(VERSION 3.20)
project(test_drive VERSION 1.0.0 LANGUAGES fortran)

list(APPEND CMAKE_MESSAGE_CONTEXT Test Drive)

<File contents>
.
.
.

list(POP_BACK CMAKE_MESSAGE_CONTEXT)
```
