### 2024-10-14 dftd3/simple-dftd3#105 @AlexBuccheri
Fortran API example requires updating in the documentation
The fortran API example given on [ReadTheDocs](https://dftd3.readthedocs.io/en/latest/api/fortran.html#performing-calculations) (set to latest) fails to compile

```shell
[ 37%] Building Fortran object CMakeFiles/test_d3.dir/main.f90.o
/Users/alexanderbuccheri/Codes/simple-dftd3/alex_test/main.f90:17:45:

   17 |    call new_rational_damping(param, inp, mol)
      |                                             1
Error: Type mismatch in argument 'self' at (1); passed CLASS(damping_param) to TYPE(rational_damping_param)
/Users/alexanderbuccheri/Codes/simple-dftd3/alex_test/main.f90:17:45:

   17 |    call new_rational_damping(param, inp, mol)

```

My CMakeLists.txt looks like:

```cmake
cmake_minimum_required(VERSION 3.20)

project(test_simple_d3 VERSION 1.0.0 LANGUAGES Fortran)

include(GNUInstallDirs)
include(FetchContent)

set("s-dftd3-url" "https://github.com/dftd3/simple-dftd3")
message(STATUS "Retrieving s-dftd3 from ${s-dftd3-url}")
FetchContent_Declare(
   "s-dftd3"
   GIT_REPOSITORY "${s-dftd3-url}"
   GIT_TAG "HEAD"
  )
FetchContent_MakeAvailable("s-dftd3")

add_executable(test_d3 main.f90)
target_link_libraries(test_d3 PUBLIC s-dftd3)
```
