### 2024-10-14 #7169 @AlexBuccheri
Great to see that the documentation has been updated and the majority of issues I opened have already been resolved.

I just tried the snippet on ReadTheDocs for fortran, but got a compilation error. See [issue 105](https://github.com/dftd3/simple-dftd3/issues/105) - I used fetchContent just to make sure it wasn't my tinkering that broke something.

I also see that the conda installation has been migrated to Mamba. Once I test this, and there's a working fortran API example in the docs, I'm happy to approve - should be doable this week. The overall software engineering in this project is high and this review has been open for too long now.
