### 2024-10-20 6d6f797
Add support for counter-poise calculations
