+++ app/driver.f90
@@ -25,15 +25,16 @@ module dftd3_app_driver
       & get_optimizedpower_damping, optimizedpower_damping_param, &
       & new_optimizedpower_damping, new_d3_model, get_pairwise_dispersion, &
       & realspace_cutoff, get_lattice_points
+   use dftd3_gcp, only : gcp_param, get_gcp_param, get_geometric_counterpoise
    use dftd3_output, only : ascii_damping_param, ascii_atomic_radii, &
       & ascii_atomic_references, ascii_system_properties, ascii_energy_atom, &
       & ascii_results, ascii_pairwise, tagged_result, json_results, &
-      & turbomole_gradient, turbomole_gradlatt
+      & turbomole_gradient, turbomole_gradlatt, ascii_gcp_param
    use dftd3_utils, only : wrap_to_central_cell
    use dftd3_citation, only : format_bibtex, is_citation_present, citation_type, &
       & get_citation, doi_dftd3_0, doi_dftd3_bj, doi_dftd3_m, doi_dftd3_op, same_citation
    use dftd3_app_help, only : header
-   use dftd3_app_cli, only : app_config, run_config, param_config, get_arguments
+   use dftd3_app_cli, only : app_config, run_config, param_config, gcp_config, get_arguments
    use dftd3_app_toml, only : param_database
    implicit none
    private
@@ -51,6 +52,8 @@ subroutine app_driver(config, error)
       call run_driver(config, error)
    type is(param_config)
       call param_driver(config, error)
+   type is(gcp_config)
+      call gcp_driver(config, error)
    end select
 end subroutine app_driver
 
@@ -62,6 +65,7 @@ subroutine run_driver(config, error)
    class(damping_param), allocatable :: param
    type(d3_param) :: inp
    type(d3_model) :: d3
+   type(gcp_param) :: gcp
    real(wp), allocatable :: energies(:), gradient(:, :), sigma(:, :)
    real(wp), allocatable :: pair_disp2(:, :), pair_disp3(:, :)
    real(wp), allocatable :: s9
@@ -179,9 +183,15 @@ subroutine run_driver(config, error)
          end block
       end if
    end if
+   if (config%gcp) then
+      call get_gcp_param(gcp, mol, config%method, config%basis)
+   end if
 
-   if (allocated(param) .and. config%verbosity > 0) then
-      call ascii_damping_param(output_unit, param, config%method)
+   if (config%verbosity > 0) then
+      if (allocated(param)) &
+         call ascii_damping_param(output_unit, param, config%method)
+      if (config%gcp) &
+         call ascii_gcp_param(output_unit, mol, gcp)
    end if
 
    if (allocated(param)) then
@@ -200,6 +210,9 @@ subroutine run_driver(config, error)
    if (allocated(param)) then
       call get_dispersion(mol, d3, param, realspace_cutoff(), energies, &
          & gradient, sigma)
+      if (config%gcp) then
+         call get_geometric_counterpoise(mol, gcp, energies, gradient, sigma)
+      end if
       energy = sum(energies)
 
       if (config%pair_resolved) then
@@ -217,50 +230,15 @@ subroutine run_driver(config, error)
          end if
       end if
       if (config%tmer) then
-         if (config%verbosity > 0) then
-            write(output_unit, '(a)') "[Info] Dispersion energy written to .EDISP"
-         end if
-         open(file=".EDISP", newunit=unit)
-         write(unit, '(f24.14)') energy
-         close(unit)
+         call tmer_writer(".EDISP", energy, "Dispersion", config%verbosity)
       end if
       if (config%grad) then
          if (allocated(config%grad_output)) then
-            open(file=config%grad_output, newunit=unit)
-            call tagged_result(unit, energy, gradient, sigma)
-            close(unit)
[…]
