+++ app/cli.f90
@@ -21,11 +21,11 @@ module dftd3_app_cli
    use dftd3, only : d3_param
    use dftd3_app_argument, only : argument_list, len
    use dftd3_app_help, only : prog_name, header, help_text, run_help_text, param_help_text, &
-      & version
+      & gcp_help_text, version
    implicit none
    private
 
-   public :: app_config, run_config, param_config, get_arguments
+   public :: app_config, run_config, param_config, gcp_config, get_arguments
 
    type, abstract :: app_config
    end type app_config
@@ -37,6 +37,8 @@ module dftd3_app_cli
       integer, allocatable :: input_format
       !> Method name
       character(len=:), allocatable :: method
+      !> Basis name
+      character(len=:), allocatable :: basis
       !> Damping paramaters
       type(d3_param) :: inp
       logical :: json = .false.
@@ -56,6 +58,7 @@ module dftd3_app_cli
       integer :: verbosity = 2
       logical :: pair_resolved = .false.
       logical :: citation = .false.
+      logical :: gcp = .false.
       character(len=:), allocatable :: citation_output
       !> Parameter data base
       character(len=:), allocatable :: db
@@ -69,6 +72,26 @@ module dftd3_app_cli
       !> Damping function
       character(len=:), allocatable :: damping
    end type param_config
+
+   type, extends(app_config) :: gcp_config
+      !> Geometry input file
+      character(len=:), allocatable :: input
+      !> Format of the geometry input
+      integer, allocatable :: input_format
+      !> Method name
+      character(len=:), allocatable :: method
+      !> Basis name
+      character(len=:), allocatable :: basis
+      logical :: json = .false.
+      character(len=:), allocatable :: json_output
+      logical :: wrap = .true.
+      logical :: tmer = .true.
+      logical :: grad = .false.
+      character(len=:), allocatable :: grad_output
+      integer :: verbosity = 2
+      logical :: citation = .false.
+      character(len=:), allocatable :: citation_output
+   end type gcp_config
 
 contains
 
@@ -131,6 +154,9 @@ subroutine get_arguments(config, error)
       case("param")
          allocate(param_config :: config)
          exit
+      case("gcp")
+         allocate(gcp_config :: config)
+         exit
       end select
    end do
    if (allocated(error)) return
@@ -146,6 +172,8 @@ subroutine get_arguments(config, error)
       call get_run_arguments(config, list, iarg, error)
    type is(param_config)
       call get_param_arguments(config, list, iarg, error)
+   type is(gcp_config)
+      call get_gcp_arguments(config, list, iarg, error)
    end select
 end subroutine get_arguments
 
@@ -266,6 +294,12 @@ subroutine get_run_arguments(config, list, start, error)
          call get_argument_as_real(arg, config%inp%s9, error)
          if (allocated(error)) exit
          config%atm = .true.
+      case("--gcp")
+         iarg = iarg + 1
+         call list%get(iarg, arg)
+         call move_alloc(arg, config%basis)
+         if (allocated(error)) exit
+         config%gcp = .true.
       case("--zero")
          config%zero = .true.
          iarg = iarg + 1
@@ -488,6 +522,115 @@ subroutine get_param_arguments(config, list, start, error)
    end if
 
 end subroutine get_param_arguments
+
+subroutine get_gcp_arguments(config, list, start, error)
+
+   !> Configuation data
+   type(gcp_config), intent(out) :: config
+
+   !> List of command line arguments
+   type(argument_list), intent(in) :: list
+
+   !> First command line argument
+   integer, intent(in) :: start
+
+   !> Error handling
+   type(error_type), allocatable, intent(out) :: error
+
+   integer :: iarg, narg
+   character(len=:), allocatable :: arg
+
+   iarg = start
[…]
