### 2024-10-20 59d03b4
Add tutorial on command line interface
