+++ doc/tutorial/first-steps-cli.rst
@@ -1,0 +1,235 @@
+Computing interaction energies
+==============================
+
+In this tutorial we want to compute the interaction energy of a non-covalently bound complex using DFT with D3 dispersion correction.
+For this purpose we are using the forth system from the S66x8 data set.\ :footcite:`golokesh`
+
+We start by providing the coordinates of the dimer system in xyz format.
+
+.. code-block:: text
+   :caption: dimer.xyz
+
+   15
+   water-peptide dimer, PBE0/def2-QZVP energy: -324.751193159385
+   O    -3.2939688    0.4402024    0.1621802 
+   H    -3.8134112    1.2387332    0.2637577 
+   H    -2.3770466    0.7564365    0.1766203 
+   C    -0.6611637   -1.4159110   -0.1449409 
+   H    -0.0112009   -2.2770229   -0.2778563 
+   H    -1.3421397   -1.3384389   -0.9888061 
+   H    -1.2741806   -1.5547070    0.7420675 
+   C     0.0935684   -0.1178981   -0.0123474 
+   O    -0.4831471    0.9573968    0.1442414 
+   N     1.4442015   -0.2154008   -0.0769653 
+   H     1.8451531   -1.1259348   -0.2064804 
+   C     2.3124436    0.9365697    0.0324778 
+   H     1.6759495    1.8048701    0.1672624 
+   H     2.9780331    0.8451145    0.8885706 
+   H     2.9069093    1.0659902   -0.8697814 
+
+Since D3 is a post SCF correction, we need to compute the SCF interaction energy with another program first.
+For this system the PBE0/def2-QZVP interaction energy is computed with Orca 5.0.4 and already included in the comment line of the xyz file.
+
+Our next step is to compute the D3 contribution to the PBE0-D3 energy, here we choose the rational damping version of D3 denoted as Becke--Johnson (BJ) damping.
+For this we call the ``s-dftd3`` command with the xyz file, here ``dimer.xyz`` and the option ``--bj`` with the functional name, here ``PBE0``.
+The output should look like:
+
+.. code-block:: text
+
+    ❯ s-dftd3 dimer.xyz --bj PBE0
+    -----------------------------------
+    s i m p l e   D F T - D 3  v1.0.0
+    -----------------------------------
+
+    Rational (Becke-Johnson) damping: PBE0-D3(BJ)
+    ---------------------
+    s6         1.0000
+    s8         1.2177
+    s9         0.0000
+    a1         0.4145
+    a2         4.8593
+    alp        14.0000
+    --------------------
+
+    Dispersion energy:      -8.2944752821052E-03 Eh
+
+    [Info] Dispersion energy written to .EDISP
+
+The output reports a few information for us, first it reports the program version we are using, second the parameters used in the dispersion calculation, we find the rational damping reported here as *PBE0-D3(BJ)*.
+Finally the dispersion energy is writtem to the screen in the output but also stored in the ``.EDISP`` file for easily accessing it without having to search the output.
+
+To compute the interaction energy for the system, we also need to compute the individual monomers.
+Since we are computing an interaction energy, the monomers are in the same geometry as in the dimer, i.e. there is no geometry change.
+For our system we copy the first three atoms and the last twelve atoms to separate files.
+
+The water monomer is given by
+
+.. code-block:: text
+   :caption: monomer-a.xyz
+
+   3
+   water monomer, PBE0/def2-QZVP energy: -76.386381675761
+   O    -3.2939688    0.4402024    0.1621802 
+   H    -3.8134112    1.2387332    0.2637577 
+   H    -2.3770466    0.7564365    0.1766203 
+
+The peptide monomer is given by
+
+.. code-block:: text
+   :caption: monomer-b.xyz
+
+   12
[…]
