### 2024-10-23 2487250
Add tutorial on command line interface (#106)
