+++ doc/tutorial/index.rst
@@ -1,0 +1,8 @@
+Tutorials
+=========
+
+This section contains tutorials on using D3.
+
+.. toctree::
+
+   first-steps-cli
