### 2024-10-23 7f2b4e6
Update CMake docs
