### 2024-10-23 0bf8f33
Update docs on downloaded dependencies
