### 2024-10-23 5d3cdb3
Update docs on build (#108)
* Update CMake docs
* Update docs on downloaded dependencies
