+++ doc/tutorial/index.rst
@@ -6,3 +6,4 @@ This section contains tutorials on using D3.
 .. toctree::
 
    first-steps-cli
+   first-steps-fortran
