### 2024-10-29 8ce13e6
Bump version to 1.2.0
