### 2024-10-29 e583332
Bump version to 1.2.0 (#112)
