+++ python/mesonpep517.toml
@@ -4,7 +4,7 @@ build-backend = "mesonpy"
 
 [project]
 name = "dftd3"
-version = "1.1.1"
+version = "1.2.0"
 description = "Python API of the DFT-D3 project"
 readme = "README.rst"
 license.text = "LGPL-3.0-or-later"
