### 2024-10-29 #7169 @awvwgk
Thanks for the patience, just released version 1.2.0 which addresses all comments from the review. Please let me know if there is anything outstanding which I overlooked.
