### 2024-10-30 #7169 @AlexBuccheri
## Feedback on v 1.2.0:

* I revisited the meson build with GCC correctly aliased - extremely fast build
* Tested the mamba installation - very smooth
* Great that you added a `requirements.txt` for sphinx builds, as I had to trial-and-error install numerous packages into my venv last time I tried it
* All in all great improvement on documenting how to build the docs

## Outstanding Issues

These are all trival, so I've just stated them here. If you'd prefer me to open Github issues I can also do that.

### README.md

I think there's a typo in:
```
pip install -f doc/requirements.txt
```
Shouldn't this be '-r' not '-f'?

------------------

When I run `python -m http.server -d _docs` I get the message:
```console
Serving HTTP on :: port 8000 (http://[::]:8000/) ...
```
I had to look this up. Apparently `[::]` is IPv6 notation for 0.0.0.0. So it _might_ be useful to say the page is then reachable at `http://0.0.0.0:8000` in the docs.

------------------

Using ` app/example.xyz` with the command line example:
```console
s-dftd3 --bj pbe0 --atm --json --noedisp --grad -- app/example.xyz
```
doesn't work. Same issue as previously reported? Rearranging so the input comes directly after `--atm`  does work:
```console
s-dftd3 --bj pbe0  --json --noedisp --grad --atm app/example.xyz
```
and it would be useful for the docs to just give this.

------------------

Any mention of cmake has disappeared from the top-level `README.md`, although it's still present
[here](https://dftd3.readthedocs.io/en/latest/installation.html). Mentioning cmake with this hyperlink would be great.

------------------

The fortran API example in the top-level `docs.md` has an erroneous line:

```fortran
call new_rational_damping(param, inp, mol)
! Should be:
call new_rational_damping(param, inp)
```

which makes it consistent with [read-the-docs](http://localhost:63343/simple-dftd3/_docs/api/fortran.html#performing-calculations)

------------------

### A Minimal Working Library Example

It would be nice to have a minimal working library call example. The tests in `app/` were quite involved, and the tutorials forced me to read more than I wanted i.e. just show me how to call the API with my DFT code 😅 . Something like this, which I cobbled together from the various examples. You may disagree, the suggestion is entirely discretionary:

  <summary>A mininal fortran example that compiles and gives an output</summary>

```fortran
program test_simple_d3
   use, intrinsic :: iso_fortran_env
   use mctc_env, only: error_type
   use mctc_io,  only: structure_type, new
   use dftd3,    only: d3_model, d3_param, rational_damping_param, &
                     & get_rational_damping, new_rational_damping, new_d3_model, &
                     & get_dispersion, realspace_cutoff
   implicit none

   character(len=:), allocatable :: method
   type(structure_type)          :: mol
   integer,          allocatable :: num(:)
   real(real64),     allocatable :: xyz(:, :)
   real(real64)                  :: energy
   real(real64),     allocatable :: gradient(:, :)
   real(real64),     allocatable :: sigma(:, :)
   type(error_type), allocatable :: error

   type(d3_model) :: disp
   type(d3_param) :: inp
   type(rational_damping_param) :: param

   method = 'dftb3'
   num = [6, 1, 1, 1, 1]
   xyz = reshape([ &
     &  0.00000000000000_real64, -0.00000000000000_real64,  0.00000000000000_real64, &
     & -1.19220800552211_real64,  1.19220800552211_real64,  1.19220800552211_real64, &
     &  1.19220800552211_real64, -1.19220800552211_real64,  1.19220800552211_real64, &
     & -1.19220800552211_real64, -1.19220800552211_real64, -1.19220800552211_real64, &
     &  1.19220800552211_real64,  1.19220800552211_real64, -1.19220800552211_real64],&
     & [3, size(num)])

   call new(mol, num, xyz, charge=0.0_real64, uhf=0)

   call get_rational_damping(inp, method, error, s9=1.0_real64)
   if (allocated(error)) return

   call new_rational_damping(param, inp)

   call new_d3_model(disp, mol)

   call get_dispersion(mol, disp, param, realspace_cutoff(), energy)  !gradient, sigma)

   write(*, *) 'Dispersion energy for ', method, ' is ', energy

end program test_simple_d3
```
