### 2024-10-30 #7169 @AlexBuccheri
@zhubonan I've given some minor suggestions on changes following the release of version 1.2.0 but otherwise I'm happy to approve this JOSS submission. The package is very clean, self-contained and well-documented. @awvwgk has taken all the feedback onboard and made improvements on what was already a polished library. Great work! 🙌🏻
