### 2024-10-30 #7169 @AlexBuccheri
> and this review has been open for too long now.

Just to clarify, I meant it would be nice for you (@awvwgk) to have the feedback and get the manuscript/library accepted. Wasn't a criticism of the time spent implementing the suggestions.
