### 2024-10-31 4c96c52
Minor updates
- implement -- to stop parsing for arguments in CLI
- fix typo for pip installation of docs requirements
- mention default URL for HTTP server
