### 2024-10-31 2697ac7
Minor updates (#113)
- implement -- to stop parsing for arguments in CLI
- fix typo for pip installation of docs requirements
- mention default URL for HTTP server
