+++ src/api.f90
@@ -1,0 +1,867 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+!> Definition of the public C-API of s-dftd3
+!>
+!>```c
+!>{!./include/s-dftd3.h!}
+!>```
+module dftd3_api
+   use iso_c_binding
+   use mctc_env, only : wp, error_type, fatal_error
+   use mctc_io_structure, only : structure_type, new
+   use dftd3_cutoff, only : realspace_cutoff
+   use dftd3_damping_mzero, only : mzero_damping_param, new_mzero_damping
+   use dftd3_damping_optimizedpower, only : optimizedpower_damping_param, &
+      & new_optimizedpower_damping
+   use dftd3_damping_rational, only : rational_damping_param, new_rational_damping
+   use dftd3_damping_zero, only : zero_damping_param, new_zero_damping
+   use dftd3_damping, only : damping_param
+   use dftd3_disp, only : get_dispersion, get_pairwise_dispersion
+   use dftd3_model, only : d3_model, new_d3_model
+   use dftd3_param, only : d3_param, get_rational_damping, get_zero_damping, &
+      & get_mrational_damping, get_mzero_damping, get_optimizedpower_damping
+   use dftd3_utils, only : wrap_to_central_cell
+   use dftd3_version, only : get_dftd3_version
+   implicit none
+   private
+
+   public :: get_version_api
+
+   public :: vp_error
+   public :: new_error_api, check_error_api, get_error_api, delete_error_api
+
+   public :: vp_structure
+   public :: new_structure_api, delete_structure_api, update_structure_api
+
+   public :: vp_model
+   public :: new_d3_model_api, delete_model_api
+
+   public :: vp_param
+   public :: new_zero_damping_api, load_zero_damping_api
+   public :: new_rational_damping_api, load_rational_damping_api
+   public :: new_mzero_damping_api, load_mzero_damping_api
+   public :: new_mrational_damping_api, load_mrational_damping_api
+   public :: new_optimizedpower_damping_api, load_optimizedpower_damping_api
+   public :: delete_param_api
+
+
+   !> Void pointer to error handle
+   type :: vp_error
+      !> Actual payload
+      type(error_type), allocatable :: ptr
+   end type vp_error
+
+   !> Void pointer to molecular structure data
+   type :: vp_structure
+      !> Actual payload
+      type(structure_type) :: ptr
+   end type vp_structure
+
+   !> Void pointer to dispersion model
+   type :: vp_model
+      !> Actual payload
+      type(d3_model) :: ptr
+      !> Additional real space cutoff
+      type(realspace_cutoff), allocatable :: cutoff
+   end type vp_model
+
+   !> Void pointer to damping parameters
[…]
