+++ src/cli.f90
@@ -1,0 +1,501 @@
+! This file is part of s-dftd3.
+! SPDX-Identifier: LGPL-3.0-or-later
+!
+! s-dftd3 is free software: you can redistribute it and/or modify it under
+! the terms of the GNU Lesser General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! s-dftd3 is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU Lesser General Public License for more details.
+!
+! You should have received a copy of the GNU Lesser General Public License
+! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
+
+module dftd3_app_cli
+   use, intrinsic :: iso_fortran_env, only : output_unit
+   use mctc_env, only : wp, error_type, fatal_error
+   use mctc_io, only : structure_type, read_structure, filetype, get_filetype
+   use dftd3, only : d3_param
+   use dftd3_app_argument, only : argument_list, len
+   use dftd3_app_help, only : prog_name, header, help_text, run_help_text, param_help_text, &
+      & version
+   implicit none
+   private
+
+   public :: app_config, run_config, param_config, get_arguments
+
+   type, abstract :: app_config
+   end type app_config
+
+   type, extends(app_config) :: run_config
+      !> Geometry input file
+      character(len=:), allocatable :: input
+      !> Format of the geometry input
+      integer, allocatable :: input_format
+      !> Method name
+      character(len=:), allocatable :: method
+      !> Damping paramaters
+      type(d3_param) :: inp
+      logical :: json = .false.
+      character(len=:), allocatable :: json_output
+      logical :: wrap = .true.
+      logical :: tmer = .true.
+      logical :: properties = .false.
+      logical :: atm = .false.
+      logical :: grad = .false.
+      character(len=:), allocatable :: grad_output
+      logical :: zero = .false.
+      logical :: rational = .false.
+      logical :: mzero = .false.
+      logical :: mrational = .false.
+      logical :: optimizedpower = .false.
+      logical :: has_param = .false.
+      integer :: verbosity = 2
+      logical :: pair_resolved = .false.
+      logical :: citation = .false.
+      character(len=:), allocatable :: citation_output
+      !> Parameter data base
+      character(len=:), allocatable :: db
+   end type run_config
+
+   type, extends(app_config) :: param_config
+      !> Data base input file
+      character(len=:), allocatable :: input
+      !> Method name
+      character(len=:), allocatable :: method
+      !> Damping function
+      character(len=:), allocatable :: damping
+   end type param_config
+
+contains
+
+
+subroutine get_argument_as_real(arg, val, error)
+   !> Index of command line argument, range [0:command_argument_count()]
+   character(len=:), intent(in), allocatable :: arg
+   !> Real value
+   real(wp), intent(out) :: val
+   !> Error handling
+   type(error_type), allocatable :: error
+
+   integer :: stat
+
+   if (.not.allocated(arg)) then
+      call fatal_error(error, "Cannot read real value, argument missing")
+      return
+   end if
+   read(arg, *, iostat=stat) val
+   if (stat /= 0) then
+      call fatal_error(error, "Cannot read real value from '"//arg//"'")
+      return
+   end if
+
+end subroutine get_argument_as_real
+
+
+subroutine get_arguments(config, error)
+
+   !> Configuation data
+   class(app_config), allocatable, intent(out) :: config
+
+   !> Error handling
+   type(error_type), allocatable, intent(out) :: error
+
+   type(argument_list) :: list
+   integer :: iarg, narg
+   character(len=:), allocatable :: arg
+
[…]
