### 2024-11-01 d122f25
Add how-to guide for documentation
