+++ doc/guide/index.rst
@@ -1,0 +1,8 @@
+How-to guides
+=============
+
+This section contains guides on how to use D3 for your applications.
+
+.. toctree::
+
+   minimal-example
