+++ doc/index.rst
@@ -8,8 +8,10 @@ The *s-dftd3* project aims to provide a user-friendly and uniform interface to t
 .. _s-dftd3: https://github.com/dftd3/simple-dftd3
 
 .. toctree::
+   :maxdepth: 2
 
    Installation <installation>
    Tutorial <tutorial/index>
+   How-to <guide/index>
    Comparison <comparison>
    API <api/index>
