+++ doc/comparison.rst
@@ -63,6 +63,8 @@ A list of projects currently using this DFT-D3 implementation is given here.
   Ring polymer molecular dynamics and rate constant calculations on black-box potential energy surfaces.\ :footcite:`steffen2023`
 `Curcuma <https://github.com/conradhuebler/curcuma>`_:
   Simple small molecular docking and conformation filtering tool.
+`MLAtom <https://github.com/dralgroup/mlatom>`_: (since 3.11.0)
+  Platform for Machine Learning-Enhanced Computational Chemistry Simulations and Workflows.\ :footcite:`dral2024`
 
 
 If your project is using *s-dftd3* feel free to add your project to this list.
