### 2024-11-05 f572877
Update C API documentation
