+++ doc/api/c.rst
@@ -7,38 +7,303 @@ The API user is required delete all objects created in the library by using the 
 
 Overall four classes of objects are provided by the library
 
-- error handlers (``dftd3_error``),
+- error handlers (:c:type:`dftd3_error`),
   used to communicate exceptional conditions and errors from the library to the user
-- structure containers (``dftd3_structure``),
+- structure containers (:c:type:`dftd3_structure`),
   used to represent the system specific information and geometry data,
   only the latter are mutable for the user
-- dispersion model objects (``dftd3_model``),
+- dispersion model objects (:c:type:`dftd3_model`),
   general model for calculating dispersion releated properties
-- damping function objects (``dftd3_param``)
+- damping function objects (:c:type:`dftd3_param`)
   polymorphic objects to represent the actual method parametrisation
 
 .. note::
 
    Generally, all quantities provided to the library are assumed to be in `atomic units <https://en.wikipedia.org/wiki/Hartree_atomic_units>`_.
 
-.. contents::
-
 
 Error handling
 --------------
 
-The library provides a light error handle type (``dftd3_error``) for storing error information
+.. c:type:: struct _dftd3_error* dftd3_error;
+
+   Error handle class
+
+The library provides a light error handle type (:c:type:`dftd3_error`) for storing error information
 The error handle requires only small overhead to construct and can only contain a single error.
 
 The handler is represented by an opaque pointer and can only be manipulated by call from the library.
 The user of those objects is required to delete the handlers again using the library provided deconstructors to avoid memory leaks.
 
+.. c:function:: dftd3_error dftd3_new_error();
+
+   :returns: New allocation for error handle
+
+   Create new error handle object
+
+.. c:function:: int dftd3_check_error(dftd3_error error);
+
+   :param error: Error handle
+   :returns: Current status of error handle, non-zero in case of error
+
+   Check error handle status
+
+.. c:function:: void dftd3_get_error(dftd3_error error, char* buffer, const int* buffersize);
+
+   :param error: Error handle
+   :param buffer: Allocation to store error message in
+   :param buffersize: Maximum length of the buffer (optional)
+
+   Get error message from error handle
+
+.. c:function:: void dftd3_delete_error(dftd3_error* error);
+
+   :param error: Error handle
+
+   Delete error handle object. The handle is set to NULL after deletion.
+
 
 Structure data
 --------------
 
+.. c:type:: struct _dftd3_structure* dftd3_structure;
+
+   Molecular structure data class
+
 The structure data is used to represent the system of interest in the library.
 It contains immutable system specific information like the number of atoms, the unique atom groups and the boundary conditions as well as mutable geometry data like cartesian coordinates and lattice parameters.
+
+.. c:function:: dftd3_structure dftd3_new_structure(dftd3_error error, const int natoms, const int* numbers, const double* positions, const double* lattice, const bool* periodic);
+
+   :param natoms: Number of atoms in the system
+   :param numbers: Atomic numbers of all atoms [natoms]
+   :param positions: Cartesian coordinates in Bohr [natoms, 3]
+   :param lattice: Lattice parameters in Bohr [3, 3] (optional)
+   :param periodic: Periodic dimension of the system [3] (optional)
+   :returns: New molecular structure data handle
+
+   Create new molecular structure data (quantities in Bohr)
+
+.. c:function:: void dftd3_update_structure(dftd3_error error, dftd3_structure mol, const double* positions, const double* lattice);
+
+   :param error: Error handle
+   :param mol: Molecular structure data handle
[…]
