### 2024-11-06 79c1365
Documentation updates (#115)
* Add reference to MLatom as user
* Update C API documentation
