+++ src/dftd3/gcp.f90
@@ -23,6 +23,7 @@ module dftd3_gcp
 
    public :: gcp_param, get_gcp_param, get_geometric_counterpoise
 
+   !> Geometric counterpoise correction
    interface get_geometric_counterpoise
       module procedure get_geometric_counterpoise
       module procedure get_geometric_counterpoise_atomic
@@ -33,11 +34,13 @@ module dftd3_gcp
 contains
 
 
+!> Geometric counterpoise correction
 subroutine get_geometric_counterpoise(mol, param, energy, gradient, sigma)
 
    !> Molecular structure data
    class(structure_type), intent(in) :: mol
 
+   !> Geometric counterpoise parameters
    type(gcp_param), intent(in) :: param
 
    !> Counter-poise energy
@@ -57,11 +60,13 @@ subroutine get_geometric_counterpoise(mol, param, energy, gradient, sigma)
 end subroutine get_geometric_counterpoise
 
 
+!> Geometric counterpoise correction with atom-resolved energies
 subroutine get_geometric_counterpoise_atomic(mol, param, energies, gradient, sigma)
 
    !> Molecular structure data
    class(structure_type), intent(in) :: mol
 
+   !> Geometric counterpoise parameters
    type(gcp_param), intent(in) :: param
 
    !> Dispersion energy
@@ -89,60 +94,59 @@ subroutine get_geometric_counterpoise_atomic(mol, param, energies, gradient, sig
 end subroutine get_geometric_counterpoise_atomic
 
 
+!> Geometric counterpoise correction
 subroutine gcp_egrad(mol, iz, emiss, slater, xv, rvdw, sigma, alpha, beta, &
-   & damp, dmp_scal, dmp_exp, ea, g)
+      & damp, dmp_scal, dmp_exp, ea, g)
+   !> Molecular structure data
    type(structure_type), intent(in) :: mol
+   !> Effective nuclear charges
    integer, intent(in) :: iz(:)
+   !> Basis set superposition error per atom
    real(wp), intent(in) :: emiss(:)
+   !> Slater exponents
    real(wp), intent(in) :: slater(:)
+   !> Number of virtual orbitals
    real(wp), intent(in) :: xv(:)
+   !> Van der Waals radii
    real(wp), intent(in) :: rvdw(:, :)
+   !> Scaling factor
    real(wp), intent(in) :: sigma
+   !> Exponential factor
    real(wp), intent(in) :: alpha
+   !> Power factor
    real(wp), intent(in) :: beta
+   !> Damping flag
    logical, intent(in) :: damp
+   !> Damping scaling factor
    real(wp), intent(in) :: dmp_scal
+   !> Damping exponent
    real(wp), intent(in) :: dmp_exp
-   real(wp), intent(out) :: ea(:)
-   real(wp), intent(out), optional :: g(:, :)
-integer n,np
-integer iat,jat,izp,jzp
-real(wp)  xyzjat(3)
-real(wp)  dum2,dum22
-real(wp)  thrR,thrE
[…]
