+++ python/dftd3/parameters.py
@@ -53,7 +53,12 @@ def get_data_file_name(base_name: str = "parameters.toml") -> str:
         dirname(__file__), "..", "..", "..", "..", "share", "s-dftd3", base_name
     )
     if not exists(data_file):
-        data_file = join(dirname(__file__), base_name)
+        # for Windows install layout
+        data_file = join(
+            dirname(__file__), "..", "..", "..", "Library", "share", "s-dftd3", base_name
+        )
+        if not exists(data_file):
+            data_file = join(dirname(__file__), base_name)
 
     return data_file
 
