### 2024-11-07 b33c556
Add additional test, update references
