+++ src/dftd3/gcp.f90
@@ -18,6 +18,7 @@ module dftd3_gcp
    use mctc_env, only : wp
    use mctc_io, only : structure_type
    use dftd3_gcp_param, only : gcp_param, get_gcp_param
+   use dftd3_cutoff, only : realspace_cutoff, get_lattice_points
    implicit none
    private
 
@@ -35,13 +36,16 @@ contains
 
 
 !> Geometric counterpoise correction
-subroutine get_geometric_counterpoise(mol, param, energy, gradient, sigma)
+subroutine get_geometric_counterpoise(mol, param, cutoff, energy, gradient, sigma)
 
    !> Molecular structure data
    class(structure_type), intent(in) :: mol
 
    !> Geometric counterpoise parameters
    type(gcp_param), intent(in) :: param
+
+   !> Realspace cutoffs
+   type(realspace_cutoff), intent(in) :: cutoff
 
    !> Counter-poise energy
    real(wp), intent(inout) :: energy
@@ -55,19 +59,22 @@ subroutine get_geometric_counterpoise(mol, param, energy, gradient, sigma)
    real(wp), allocatable :: energies(:)
 
    allocate(energies(mol%nat), source=0.0_wp)
-   call get_geometric_counterpoise_atomic(mol, param, energies, gradient, sigma)
+   call get_geometric_counterpoise_atomic(mol, param, cutoff, energies, gradient, sigma)
    energy = energy + sum(energies)
 end subroutine get_geometric_counterpoise
 
 
 !> Geometric counterpoise correction with atom-resolved energies
-subroutine get_geometric_counterpoise_atomic(mol, param, energies, gradient, sigma)
+subroutine get_geometric_counterpoise_atomic(mol, param, cutoff, energies, gradient, sigma)
 
    !> Molecular structure data
    class(structure_type), intent(in) :: mol
 
    !> Geometric counterpoise parameters
    type(gcp_param), intent(in) :: param
+
+   !> Realspace cutoffs
+   type(realspace_cutoff), intent(in) :: cutoff
 
    !> Dispersion energy
    real(wp), intent(inout) :: energies(:)
@@ -78,27 +85,40 @@ subroutine get_geometric_counterpoise_atomic(mol, param, energies, gradient, sig
    !> Dispersion virial
    real(wp), intent(inout), optional :: sigma(:, :)
 
+   real(wp), allocatable :: lattr(:, :)
+
    if (allocated(param%emiss)) then
-      call gcp_egrad(mol, param%zeff, param%emiss, param%slater, param%xv, param%rvdw, &
-         & param%sigma, param%alpha, param%beta, param%damp, param%dmp_scal, param%dmp_exp, &
-         & energies, gradient)
+      call get_lattice_points(mol%periodic, mol%lattice, cutoff%gcp, lattr)
+      call gcp_egrad(mol, lattr, cutoff%gcp, param%zeff, param%emiss, param%slater, &
+         & param%xv, param%rvdw, param%sigma, param%alpha, param%beta, param%damp, &
+         & param%dmp_scal, param%dmp_exp, energies, gradient, sigma)
+   end if
+
+   if (param%srb .or. param%base) then
[…]
