+++ src/dftd3/gcp/param.f90
@@ -23,22 +23,39 @@ module dftd3_gcp_param
 
    public :: gcp_param, get_gcp_param
 
+   !> Parameters for the geometric counter-poise correction
    type :: gcp_param
+      !> Basis set superposition error correction
       real(wp), allocatable :: emiss(:)
+      !> Number of virtual orbitals
       real(wp), allocatable :: xv(:)
+      !> Slater exponents
       real(wp), allocatable :: slater(:)
+      !> Van der Waals radii for effective nuclear charges
       real(wp), allocatable :: rvdw(:, :)
+      !> Van der Waals radii for true nuclear charges
       real(wp), allocatable :: rvdw_srb(:, :)
+      !> Effective nuclear charges
       integer, allocatable :: zeff(:)
+      !> Scaling factor for the counter-poise correction
       real(wp) :: sigma = 0.0_wp
+      !> Exponential parameter
       real(wp) :: alpha = 0.0_wp
+      !> Power parameter
       real(wp) :: beta = 0.0_wp
+      !> Damping enabled
       logical :: damp = .false.
+      !> Damping scaling factor
       real(wp) :: dmp_scal = 4.0_wp
+      !> Damping exponent
       real(wp) :: dmp_exp = 6.0_wp
+      !> Short-range bond correction
       logical :: srb = .false.
+      !> Short-range bond correction radii scaling factor
       real(wp) :: rscal = 0.0_wp
+      !> Short-range bond correction scaling factor
       real(wp) :: qscal = 0.0_wp
+      !> Short-range bond correction for HF-3c
       logical :: base = .false.
     end type
 
@@ -57,12 +74,12 @@ module dftd3_gcp_param
          p_ccdz_bas, &
          p_accdz_bas, &
          p_pobtz_bas, &
-         p_pobdzvp_bas, &
+         ! p_pobdzvp_bas, &
          p_minix_bas, &
          p_gcore_bas, &
          p_2g_bas, &
          p_dzp_bas, &
-         p_hsv_bas, &
+         ! p_hsv_bas, &
          p_dz_bas, &
          p_msvp_bas, &
          p_lanl2_bas, &
@@ -473,13 +490,14 @@ module dftd3_gcp_param
 
 contains
 
-subroutine get_gcp_param(param, mol, method, basis)
+subroutine get_gcp_param(param, mol, method, basis, eta)
    type(gcp_param), intent(out) :: param
    type(structure_type), intent(in) :: mol
    character(len=*), intent(in), optional :: method
    character(len=*), intent(in), optional :: basis
+   real(wp), intent(in), optional :: eta
 
-   real(wp) :: eta, eta_spec
+   real(wp) :: eta_, eta_spec
    real(wp), allocatable :: nbas(:)
    integer :: isp, izp, jsp, basis_id, method_id
    integer, allocatable :: nel(:)
@@ -495,7 +513,8 @@ subroutine get_gcp_param(param, mol, method, basis)
    if (present(basis)) basis_id = get_basis_id(basis)
    if (basis_id == p_unknown_bas .and. method_id == p_unknown_method) return
 
-   eta = 0.0_wp
+   eta_ = 0.0_wp
+   if (present(eta)) eta_ = eta
    eta_spec = 0.0_wp
 
    allocate(param%zeff(mol%nid))
@@ -538,17 +557,17 @@ subroutine get_gcp_param(param, mol, method, basis)
       select case(method_id)
       case (p_hf_method) ! RMS=0.3218975
[…]

[…]
