+++ config/cmake/Findmctc-lib.cmake
@@ -16,8 +16,8 @@
 
 set(_lib "mctc-lib")
 set(_pkg "MCTCLIB")
-set(_url "https://github.com/grimme-lab/mctc-lib")
-set(_rev "v0.3.0")
+set(_url "https://github.com/thfroitzheim/mctc-lib")
+set(_rev "ncoord")
 
 if(NOT DEFINED "${_pkg}_FIND_METHOD")
   if(DEFINED "${PROJECT_NAME}-dependency-method")
