### 2024-11-12 d9ab9eb
Move ncoord and data to mctc-lib
