+++ app/driver.f90
@@ -18,7 +18,8 @@ module dftd3_app_driver
    use, intrinsic :: iso_fortran_env, only : output_unit, input_unit
    use mctc_env, only : wp, error_type, fatal_error
    use mctc_io, only : structure_type, read_structure, filetype, get_filetype
-   use dftd3, only : damping_param, d3_param, d3_model, get_coordination_number, &
+   use mctc_ncoord, only : ncoord_type, new_ncoord
+   use dftd3, only : damping_param, d3_param, d3_model, &
       & get_dispersion, get_zero_damping, zero_damping_param, new_zero_damping, &
       & get_rational_damping, rational_damping_param, new_rational_damping, &
       & get_mzero_damping, mzero_damping_param, new_mzero_damping, get_mrational_damping, &
@@ -289,6 +290,7 @@ subroutine property_calc(unit, mol, disp, verbosity)
 
    integer :: mref
    real(wp), allocatable :: cn(:), gwvec(:, :), c6(:, :), lattr(:, :)
+   class(ncoord_type), allocatable :: ncoord
 
    if (verbosity > 1) then
       call ascii_atomic_radii(unit, mol, disp)
@@ -299,8 +301,9 @@ subroutine property_calc(unit, mol, disp, verbosity)
 
    mref = maxval(disp%ref)
    allocate(cn(mol%nat), gwvec(mref, mol%nat), c6(mol%nat, mol%nat))
+   call new_ncoord(ncoord, mol, "exp", cutoff=30.0_wp, rcov=disp%rcov)
    call get_lattice_points(mol%periodic, mol%lattice, 30.0_wp, lattr)
-   call get_coordination_number(mol, lattr, 30.0_wp, disp%rcov, cn)
+   call ncoord%get_coordination_number(mol, lattr, cn)
    call disp%weight_references(mol, cn, gwvec)
    call disp%get_atomic_c6(mol, gwvec, c6=c6)
 
