+++ src/dftd3/ncoord.f90
@@ -1,295 +1,0 @@
-! This file is part of s-dftd3.
-! SPDX-Identifier: LGPL-3.0-or-later
-!
-! s-dftd3 is free software: you can redistribute it and/or modify it under
-! the terms of the GNU Lesser General Public License as published by
-! the Free Software Foundation, either version 3 of the License, or
-! (at your option) any later version.
-!
-! s-dftd3 is distributed in the hope that it will be useful,
-! but WITHOUT ANY WARRANTY; without even the implied warranty of
-! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-! GNU Lesser General Public License for more details.
-!
-! You should have received a copy of the GNU Lesser General Public License
-! along with s-dftd3.  If not, see <https://www.gnu.org/licenses/>.
-
-module dftd3_ncoord
-   use mctc_env, only : wp
-   use mctc_io, only : structure_type
-   implicit none
-   private
-
-   public :: get_coordination_number, add_coordination_number_derivs
-
-
-   !> Steepness of counting function
-   real(wp), parameter :: kcn = 16.0_wp
-
-
-contains
-
-
-!> Geometric fractional coordination number, supports exponential counting
-!> functions.
-subroutine get_coordination_number(mol, trans, cutoff, rcov, cn, dcndr, dcndL)
-
-   !> Molecular structure data
-   type(structure_type), intent(in) :: mol
-
-   !> Lattice points
-   real(wp), intent(in) :: trans(:, :)
-
-   !> Real space cutoff
-   real(wp), intent(in) :: cutoff
-
-   !> Covalent radius
-   real(wp), intent(in) :: rcov(:)
-
-   !> Error function coordination number.
-   real(wp), intent(out) :: cn(:)
-
-   !> Derivative of the CN with respect to the Cartesian coordinates.
-   real(wp), intent(out), optional :: dcndr(:, :, :)
-
-   !> Derivative of the CN with respect to strain deformations.
-   real(wp), intent(out), optional :: dcndL(:, :, :)
-
-   if (present(dcndr) .and. present(dcndL)) then
-      call ncoord_dexp(mol, trans, cutoff, rcov, cn, dcndr, dcndL)
-   else
-      call ncoord_exp(mol, trans, cutoff, rcov, cn)
-   end if
-
-end subroutine get_coordination_number
-
-
-subroutine ncoord_exp(mol, trans, cutoff, rcov, cn)
-
-   !> Molecular structure data
-   type(structure_type), intent(in) :: mol
-
-   !> Lattice points
-   real(wp), intent(in) :: trans(:, :)
-
-   !> Real space cutoff
-   real(wp), intent(in) :: cutoff
-
-   !> Covalent radius
-   real(wp), intent(in) :: rcov(:)
-
-   !> Error function coordination number.
-   real(wp), intent(out) :: cn(:)
-
-   integer :: iat, jat, izp, jzp, itr
-   real(wp) :: r2, r1, rc, rij(3), countf, cutoff2
-
-   cn(:) = 0.0_wp
-   cutoff2 = cutoff**2
-
-   !$omp parallel do schedule(runtime) default(none) reduction(+:cn) &
-   !$omp shared(mol, trans, cutoff2, rcov) &
-   !$omp private(jat, itr, izp, jzp, r2, rij, r1, rc, countf)
-   do iat = 1, mol%nat
-      izp = mol%id(iat)
-      do jat = 1, iat
-         jzp = mol%id(jat)
-
[…]
