### 2024-11-12 08881fc
Move ncoord and data to mctc-lib
