### 2024-11-15 9f5a3b6
Bump version to 1.2.1
