+++ python/dftd3/__init__.py
@@ -18,4 +18,4 @@
 # make sure we have a CFFI available
 import cffi
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
