### 2024-11-15 3425201
Release version 1.2.1 (#118)
