### 2024-11-15 #7169 @awvwgk
Just released as version 1.2.1 with the latest changes from the review included and updated the paper accordingly.
