+++ doc/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
   - name: Microsoft Research, AI for Science, The Netherlands
     index: 1
-date: 15 November 2024
+date: 27 July 2024
 bibliography: _static/references.bib
 ---
 
