### 2024-11-26 #7169 @awvwgk
> Great, could you please also update the title and the license of the Zenodo archive to match that of the paper, and also also add your ORCID to the archive's author list. There is no need to make a new archive - you should be able to edit the metadata of the current one. Thanks

The zenodo archive has been updated.
