### 2024-11-27 #7169 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
