### 2024-11-28 82f1919
Add periodicity for short range bound correction
