+++ src/dftd3/gcp.f90
@@ -86,12 +86,22 @@ subroutine get_geometric_counterpoise_atomic(mol, param, cutoff, energies, gradi
    real(wp), intent(inout), optional :: sigma(:, :)
 
    real(wp), allocatable :: lattr(:, :)
+   real(wp), parameter :: rexp_base = 0.75_wp, zexp_base = 1.5_wp, rexp_srb = -1.0_wp, zexp_srb = 0.5_wp
+   logical :: grad
+
+   grad = present(gradient) .and. present(sigma)
 
    if (allocated(param%emiss) .and. allocated(param%slater) .and. allocated(param%xv)) then
       call get_lattice_points(mol%periodic, mol%lattice, cutoff%gcp, lattr)
-      call gcp_egrad(mol, lattr, cutoff%gcp, param%zeff, param%emiss, param%slater, &
-         & param%xv, param%rvdw, param%sigma, param%alpha, param%beta, param%damp, &
-         & param%dmp_scal, param%dmp_exp, energies, gradient, sigma)
+      if (grad) then
+         call gcp_deriv(mol, lattr, cutoff%gcp, param%zeff, param%emiss, param%slater, &
+            & param%xv, param%rvdw, param%sigma, param%alpha, param%beta, param%damp, &
+            & param%dmp_scal, param%dmp_exp, energies, gradient, sigma)
+      else
+         call gcp_energy(mol, lattr, cutoff%gcp, param%zeff, param%emiss, param%slater, &
+            & param%xv, param%rvdw, param%sigma, param%alpha, param%beta, param%damp, &
+            & param%dmp_scal, param%dmp_exp, energies)
+      end if
    end if
 
    if (param%srb .or. param%base) then
@@ -99,19 +109,114 @@ subroutine get_geometric_counterpoise_atomic(mol, param, cutoff, energies, gradi
    end if
 
    if (param%srb) then
-      call srb_egrad2(mol, lattr, cutoff%srb, mol%num, param%rvdw_srb, param%rscal, param%qscal, &
-         & energies, gradient, sigma)
+      if (grad) then
+         call srb_deriv(mol, lattr, cutoff%srb, mol%num, param%rvdw_srb, param%rscal, param%qscal, &
+            & rexp_srb, zexp_srb, energies, gradient, sigma)
+      else
+         call srb_energy(mol, lattr, cutoff%srb, mol%num, param%rvdw_srb, param%rscal, param%qscal, &
+            & rexp_srb, zexp_srb, energies)
+      end if
    end if
 
    if (param%base) then
-      call basegrad(mol, lattr, cutoff%srb, param%zeff, param%rvdw, param%rscal, param%qscal, &
-         & energies, gradient, sigma)
+      if (grad) then
+         call srb_deriv(mol, lattr, cutoff%srb, param%zeff, param%rvdw, param%rscal, param%qscal, &
+            & rexp_base, zexp_base, energies, gradient, sigma)
+      else
+         call srb_energy(mol, lattr, cutoff%srb, param%zeff, param%rvdw, param%rscal, param%qscal, &
+            & rexp_base, zexp_base, energies)
+      end if
    end if
 end subroutine get_geometric_counterpoise_atomic
 
 
 !> Geometric counterpoise correction
-subroutine gcp_egrad(mol, trans, cutoff, iz, emiss, slater, xv, rvdw, escal, alpha, beta, &
[…]
