### 2024-11-29 #7169 @AlexBuccheri
> Thanks to everyone involved, @Kevin-Mattheus-Moerman, @zhubonan, @shahmoradi, @AlexBuccheri.

Great to see this get submission get wrapped up. I'm looking forward to adding simple-D3 support for Octopus in the near-future.
